wip(iter-247): split hints.rs (5059 lines) into hints/ submodules
