+++ crates/hyalo-cli/src/hints.rs
@@ -455,4605 +455,46 @@ fn files_from_hints(counters: Option<FilesFromCounterSummary>) -> Vec<Hint> {
 }
 
 // ---------------------------------------------------------------------------
-// Command builder
+// Submodules (iteration 247)
 // ---------------------------------------------------------------------------
-
-/// Push the global flags that were explicitly passed on the CLI.
-///
-/// Always the **last** thing a hint builder pushes (iter-213, dogfood UX-5).
-/// Several builders used to push globals mid-command and then append
-/// `--glob`/file targets after them, so the same `--format text` landed in a
-/// different position depending on which hint you were reading — the block of
-/// hints under one result set did not look like variations of one command.
-fn push_global_flags(parts: &mut Vec<String>, ctx: &HintContext) {
-    if let Some(dir) = &ctx.dir {
-        parts.push("--dir".to_owned());
-        parts.push(shell_quote(dir));
-    }
-    if let Some(fmt) = &ctx.format {
-        parts.push("--format".to_owned());
-        parts.push(shell_quote(fmt));
-    }
-    if ctx.hints {
-        parts.push("--hints".to_owned());
-    }
-}
-
-/// Push the graph/title filters that scope a `find` query (`--broken-links`,
-/// `--orphan`, `--dead-end`, `--title`) so derived hints reproduce the same
-/// filtered set. Kept separate from `push_global_flags` because these are
-/// `find`-specific, not global, flags. `--sort`/`--reverse` are pushed
-/// separately by [`push_find_sort`] since a couple of derived hints (e.g. the
-/// literal "Sort by most recently modified" suggestion) intentionally
-/// override them rather than preserve the active query's ordering.
-fn push_find_graph_filters(parts: &mut Vec<String>, ctx: &HintContext) {
-    if ctx.broken_links_filter {
-        parts.push("--broken-links".to_owned());
-    }
-    if ctx.orphan_filter {
-        parts.push("--orphan".to_owned());
-    }
-    if ctx.dead_end_filter {
-        parts.push("--dead-end".to_owned());
-    }
-    if let Some(title) = &ctx.title_filter {
-        parts.push("--title".to_owned());
-        parts.push(shell_quote(title));
-    }
-}
-
-/// Push the active `--sort`/`--reverse` so derived hints (show-all, narrow-by-tag,
-/// filter-by-status) reproduce the same ordering as the query that produced
-/// them — otherwise a truncated, sorted result set's "Show all" hint would
-/// silently revert to default ordering, changing which rows lead the output
-/// even though the *count* still matches (a milder variant of BUG-8).
-fn push_find_sort(parts: &mut Vec<String>, ctx: &HintContext) {
-    if let Some(sort) = &ctx.sort {
-        parts.push("--sort".to_owned());
-        parts.push(shell_quote(sort));
-        if ctx.reverse {
-            parts.push("--reverse".to_owned());
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// HintBuilder (ARCH-4, iter-225)
-// ---------------------------------------------------------------------------
-
-/// Argv-based builder for the `cmd` string of a [`Hint`].
-///
-/// Before iter-225, hint commands were hand-assembled `String`s — parallel
-/// copies of the CLI surface that could (and once did) reference flags the
-/// real command does not accept (`tags --limit 0`), which is exactly why
-/// `tests/e2e/hint_execution.rs` exists. `HintBuilder` is the single path
-/// forward: the command is assembled as an *argv vector* and serialized
-/// through [`shell_quote`], and [`HintBuilder::argv`] exposes that vector so
-/// tests can feed it straight back into the real clap parser
-/// (`crate::cli::args::Cli::try_parse_from`) — a hinted command that would
-/// not run is now a unit-test failure, not an e2e-spawn discovery.
-///
-/// All new/edited hints must go through this API; the guard test
-/// `no_raw_hyalo_command_literals` in this file fails the suite when a new
-/// hand-written `"hyalo …"` string literal appears in non-test source.
-#[derive(Debug, Clone)]
-pub struct HintBuilder {
-    parts: Vec<String>,
-}
-
-impl HintBuilder {
-    /// Start a command: `hyalo <subcommand>` (or a subcommand group like
-    /// `hyalo task toggle` — pass the words one by one via [`Self::raw`]).
-    #[must_use]
-    pub fn cmd(subcommand: &str) -> Self {
-        let mut parts = vec!["hyalo".to_owned()];
-        parts.extend(subcommand.split_whitespace().map(str::to_owned));
-        Self { parts }
-    }
-
-    /// Append one shell-quoted argument (paths, patterns, values).
-    #[must_use]
-    pub fn arg(mut self, arg: &str) -> Self {
-        self.parts.push(shell_quote(arg));
-        self
-    }
-
-    /// Append pre-formed, unquoted tokens (flags like `--dry-run`, already
-    /// validated enum values). Use sparingly — prefer [`Self::flag`] and
-    /// [`Self::flag_value`].
-    #[must_use]
-    pub fn raw(mut self, token: &str) -> Self {
-        self.parts.push(token.to_owned());
-        self
-    }
-
-    /// Append a bare flag, e.g. `--apply`.
-    #[must_use]
-    pub fn flag(self, flag: &str) -> Self {
-        self.raw(flag)
-    }
-
-    /// Append a flag and its shell-quoted value, e.g. `--rule HYALO006`.
-    #[must_use]
-    pub fn flag_value(self, flag: &str, value: &str) -> Self {
-        self.raw(flag).arg(value)
-    }
-
-    /// Append multiple shell-quoted arguments at once.
-    #[must_use]
-    pub fn args<I, S>(mut self, args: I) -> Self
-    where
-        I: IntoIterator<Item = S>,
-        S: AsRef<str>,
-    {
-        for a in args {
-            self.parts.push(shell_quote(a.as_ref()));
-        }
-        self
-    }
-
-    /// The raw argv vector (including the leading `"hyalo"`), for feeding
-    /// into the real clap parser in tests.
-    #[must_use]
-    pub fn argv(&self) -> &[String] {
-        &self.parts
-    }
-
-    /// Serialize to the display string: space-joined, each value already
-    /// shell-quoted by [`shell_quote`].
-    #[must_use]
-    pub fn build(&self) -> String {
-        self.parts.join(" ")
-    }
-
-    // --- crate-internal plumbing used by the `build_command_*` family ---
-    // The family predates the builder and composes shared `push_*` helpers
-    // over a `Vec<String>`; these accessors let it migrate without
-    // duplicating the argv/serialisation rules (ARCH-4, iter-225).
-
-    /// A builder holding only the `hyalo` program token.
-    pub(crate) fn empty() -> Self {
-        Self {
-            parts: vec!["hyalo".to_owned()],
-        }
-    }
-
-    /// Append one raw (unquoted) token in place.
-    pub(crate) fn push_raw(&mut self, token: &str) {
-        self.parts.push(token.to_owned());
-    }
-
-    /// Append one shell-quoted token in place.
-    pub(crate) fn push_quoted(&mut self, token: &str) {
-        self.parts.push(shell_quote(token));
-    }
-
-    /// Append the context's global flags (`--dir`, `--format`, `--hints`) and
-    /// serialize. The last thing every derived hint does (iter-213, UX-5).
-    pub(crate) fn finish(mut self, ctx: &HintContext) -> String {
-        push_global_flags(&mut self.parts, ctx);
-        self.build()
-    }
-}
-
-/// Build a command that intentionally omits `--glob` (for file-specific hints).
-fn build_command_no_glob(ctx: &HintContext, args: &[&str]) -> String {
-    let mut b = HintBuilder::empty();
-    for arg in args {
-        b.push_quoted(arg);
-    }
-    b.finish(ctx)
-}
-
-/// Build a command where `file_arg` is a positional file path following `subcommand_args`.
-///
-/// If `file_arg` starts with `-`, emits `--file <path>` instead of the bare positional
-/// to prevent clap from interpreting the filename as a flag.
-fn build_command_with_file(
-    ctx: &HintContext,
-    subcommand_args: &[&str],
-    file_arg: &str,
-    trailing_args: &[&str],
-) -> String {
-    let mut b = HintBuilder::empty();
-    for arg in subcommand_args {
-        b.push_quoted(arg);
-    }
-    push_file_positional(&mut b.parts, file_arg);
-    for arg in trailing_args {
-        b.push_quoted(arg);
-    }
-    b.finish(ctx)
-}
-
-/// Build a command that propagates `--glob` when present.
-fn build_command_with_glob(ctx: &HintContext, args: &[&str]) -> String {
-    let mut b = HintBuilder::empty();
-    for arg in args {
-        b.push_quoted(arg);
-    }
-    for glob in &ctx.glob {
-        b.push_raw("--glob");
-        b.push_quoted(glob);
-    }
-    b.finish(ctx)
-}
-
-/// Like `build_command_with_glob` but also preserves `--file` / positional file
-/// targets so that lint hints don't widen scope from a single file to the whole
-/// vault.
-fn build_command_with_glob_and_files(ctx: &HintContext, args: &[&str]) -> String {
-    let mut b = HintBuilder::empty();
-    for arg in args {
-        b.push_quoted(arg);
-    }
-    for glob in &ctx.glob {
-        b.push_raw("--glob");
-        b.push_quoted(glob);
-    }
-    for ft in &ctx.file_targets {
-        b.push_quoted(ft);
-    }
-    b.finish(ctx)
-}
-
-/// Build a `find` command that preserves the caller's existing filters (property,
-/// tag, task, file targets) plus `--glob`, then appends `extra_args`.  Use this for
-/// hints like sort and limit that refine the current query without changing its scope.
-fn build_find_command_preserving_filters(ctx: &HintContext, extra_args: &[&str]) -> String {
-    let mut b = HintBuilder::cmd("find");
-    for pf in &ctx.property_filters {
-        b.push_raw("--property");
-        b.push_quoted(pf);
-    }
-    for tf in &ctx.tag_filters {
-        b.push_raw("--tag");
-        b.push_quoted(tf);
-    }
-    if let Some(task) = &ctx.task_filter {
-        b.push_raw("--task");
-        b.push_quoted(task);
-    }
-    for ft in &ctx.file_targets {
-        b.push_raw("--file");
-        b.push_quoted(ft);
-    }
-    push_find_graph_filters(&mut b.parts, ctx);
-    push_find_sort(&mut b.parts, ctx);
-    for arg in extra_args {
-        b.push_quoted(arg);
-    }
-    push_find_index_file(&mut b.parts, ctx);
-    for glob in &ctx.glob {
-        b.push_raw("--glob");
-        b.push_quoted(glob);
-    }
-    b.finish(ctx)
-}
-
-/// Render a snapshot-index path for a hint command, preferring the shortest
-/// spelling that still runs verbatim from the user's working directory.
-///
-/// A snapshot path is the longest single token any hint carries, and a `find`
-/// listing repeats the *same* one on every derived query — four or five copies
-/// of one absolute path in a five-line block (iter-208a / UX-5). Eliding it
-/// from the repeats is not an option: hints have to stay copy-pasteable, and a
-/// `find` hint that quietly loses `--index-file` rescans the vault and answers
-/// a different question. Shortening the path to its working-directory-relative
-/// form keeps every hint runnable while removing most of the bulk, since an
-/// index almost always lives inside the project it indexes.
-#[must_use]
-pub fn shorten_index_path_for_hint(path: &std::path::Path) -> String {
-    let absolute = path.display().to_string();
-    let Ok(cwd) = std::env::current_dir() else {
-        return absolute;
-    };
-    match path.strip_prefix(&cwd) {
-        Ok(rel) if !rel.as_os_str().is_empty() => {
-            let relative = rel.to_string_lossy().replace('\\', "/");
-            if relative.len() < absolute.len() {
-                relative
-            } else {
-                absolute
-            }
-        }
-        _ => absolute,
-    }
-}
-
-/// Push `--index-file <path>` when the query ran against an explicit non-default
-/// snapshot index. Derived `find` hints must query the same index or they would
-/// silently rescan the vault (BUG-7 audit: `--index-file` was a dropped flag).
-fn push_find_index_file(parts: &mut Vec<String>, ctx: &HintContext) {
-    match &ctx.find_index {
-        FindIndexHint::None => {}
-        FindIndexHint::Default => parts.push("--index".to_owned()),
-        FindIndexHint::File(path) => {
-            parts.push("--index-file".to_owned());
-            parts.push(shell_quote(path));
-        }
-    }
-}
-
-/// Build a `find` command that preserves every active filter (property, tag,
-/// task, file, graph, title, glob, index-file) *and* appends the caller's
-/// existing body-search pattern as the leading positional argument. Used by
-/// derived hints (narrow-by-tag / filter-by-status) that must compose with the
-/// current query rather than replace it.
-fn build_find_command_composing(ctx: &HintContext, extra_args: &[&str]) -> String {
-    let mut b = HintBuilder::cmd("find");
-    if let Some(pat) = &ctx.body_pattern {
-        b.push_quoted(pat);
-    }
-    for pf in &ctx.property_filters {
-        b.push_raw("--property");
-        b.push_quoted(pf);
-    }
-    for tf in &ctx.tag_filters {
-        b.push_raw("--tag");
-        b.push_quoted(tf);
-    }
-    if let Some(task) = &ctx.task_filter {
-        b.push_raw("--task");
-        b.push_quoted(task);
-    }
-    for ft in &ctx.file_targets {
-        b.push_raw("--file");
-        b.push_quoted(ft);
-    }
-    push_find_graph_filters(&mut b.parts, ctx);
-    push_find_sort(&mut b.parts, ctx);
-    for arg in extra_args {
-        b.push_quoted(arg);
-    }
-    push_find_index_file(&mut b.parts, ctx);
-    for glob in &ctx.glob {
-        b.push_raw("--glob");
-        b.push_quoted(glob);
-    }
-    b.finish(ctx)
-}
-
-/// Build a `find` command that replaces the body search pattern with `new_pattern`
-/// while preserving all other existing filters (property, tag, task, file targets,
-/// glob). The pattern is inserted as a positional argument immediately after `find`.
-fn build_find_command_with_pattern(ctx: &HintContext, new_pattern: &str) -> String {
-    let mut b = HintBuilder::cmd("find");
-    b.push_quoted(new_pattern);
-    for pf in &ctx.property_filters {
-        b.push_raw("--property");
-        b.push_quoted(pf);
-    }
-    for tf in &ctx.tag_filters {
-        b.push_raw("--tag");
-        b.push_quoted(tf);
-    }
-    if let Some(task) = &ctx.task_filter {
-        b.push_raw("--task");
-        b.push_quoted(task);
-    }
-    for ft in &ctx.file_targets {
-        b.push_raw("--file");
-        b.push_quoted(ft);
-    }
-    for glob in &ctx.glob {
-        b.push_raw("--glob");
-        b.push_quoted(glob);
-    }
-    b.finish(ctx)
-}
-
-/// Push a file argument that is safe as a positional arg.
-///
-/// If the filename starts with `-`, clap would interpret it as a flag.
-/// In that case, emit `--file <path>` (flag form) instead of the bare positional.
-fn push_file_positional(parts: &mut Vec<String>, file: &str) {
-    if file.starts_with('-') {
-        parts.push("--file".to_owned());
-        parts.push(shell_quote(file));
-    } else {
-        parts.push(shell_quote(file));
-    }
-}
-
-/// Wrap a string in single-quotes if it contains any shell-special characters.
-///
-/// Uses an allowlist of safe characters — anything not in the list triggers quoting.
-/// Single-quoting avoids variable expansion and is safer than double-quoting.
-pub fn shell_quote(s: &str) -> String {
-    if s.is_empty()
-        || s.chars().any(|c| {
-            !matches!(c, 'a'..='z' | 'A'..='Z' | '0'..='9' | '-' | '_' | '.' | '/' | ':' | '@' | '=' | ',' | '+')
-        })
-    {
-        // In single-quoted strings, the only character that needs escaping is '
-        // which is done by ending the quote, adding an escaped quote, and reopening.
-        format!("'{}'", s.replace('\'', "'\\''"))
-    } else {
-        s.to_owned()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Status priority helpers
-// ---------------------------------------------------------------------------
-
-/// Priority rank for a status value: lower = more interesting.
-fn status_priority(value: &str) -> u8 {
-    if value.eq_ignore_ascii_case("in-progress")
-        || value.eq_ignore_ascii_case("in progress")
-        || value.eq_ignore_ascii_case("active")
-    {
-        0
-    } else if value.eq_ignore_ascii_case("planned") || value.eq_ignore_ascii_case("todo") {
-        1
-    } else if value.eq_ignore_ascii_case("draft") || value.eq_ignore_ascii_case("idea") {
-        2
-    } else if value.eq_ignore_ascii_case("completed")
-        || value.eq_ignore_ascii_case("done")
-        || value.eq_ignore_ascii_case("archived")
-    {
-        4
-    } else {
-        3
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Shared helpers
-// ---------------------------------------------------------------------------
-
-/// Extract the first modified file path from mutation output (single object or array).
-fn first_modified_file(data: &serde_json::Value) -> Option<&str> {
-    fn extract(obj: &serde_json::Value) -> Option<&str> {
-        obj.get("modified")
-            .and_then(|m| m.as_array())
-            .and_then(|a| a.first())
-            .and_then(|f| f.as_str())
-    }
-    if let Some(arr) = data.as_array() {
-        arr.iter().find_map(extract)
-    } else {
-        extract(data)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Per-source hint generators
-// ---------------------------------------------------------------------------
-
-fn hints_for_summary(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    // Compute the large-vault create-index hint upfront so it can be
-    // prepended at the end. This guarantees the hint is visible even when all
-    // MAX_HINTS slots would otherwise be consumed by orphan / broken-link /
-    // links-fix hints on large vaults (NEW-1).
-    let create_index_hint: Option<Hint> = if !ctx.has_index && !ctx.quiet {
-        let files_total = data
-            .get("files")
-            .and_then(|f| f.get("total"))
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        if files_total > LARGE_VAULT_FILE_COUNT {
-            Some(Hint::new(
-                format!("Vault has {files_total} files — create an index for faster queries"),
-                build_command_no_glob(ctx, &["create-index"]),
-            ))
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    let mut hints = Vec::new();
-
-    hints.push(Hint::new(
-        "Browse property names and types",
-        build_command_with_glob(ctx, &["properties"]),
-    ));
-    hints.push(Hint::new(
-        "Browse tags and their counts",
-        build_command_with_glob(ctx, &["tags"]),
-    ));
-
-    // Suggest lint early when there are schema violations — high priority so it
-    // is not pushed out by orphans/dead-ends/broken-links hints.
-    if let Some(schema_obj) = data.get("schema") {
-        let errors = schema_obj
-            .get("errors")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        let warnings = schema_obj
-            .get("warnings")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        if (errors > 0 || warnings > 0) && hints.len() < MAX_HINTS {
-            let errors_label = if errors == 1 { "error" } else { "errors" };
-            let warns_label = if warnings == 1 { "warning" } else { "warnings" };
-            // The summary counter measures SCHEMA (frontmatter) violations only
-            // — it does not include MD body rules that plain `hyalo lint` runs.
-            // Label it "Schema" and point at `hyalo lint --rule SCHEMA` so the
-            // hinted command reproduces these exact counts. Bare `hyalo lint`
-            // reported wildly different totals (BUG-9: 5 errors / 12 warnings in
-            // the summary vs 0 / 660 from `hyalo lint`); the counter now also
-            // applies `[lint] ignore` globs, matching what `--rule SCHEMA` sees.
-            hints.push(Hint::new(
-                format!("Schema: {errors} {errors_label}, {warnings} {warns_label}"),
-                build_command_with_glob(ctx, &["lint", "--rule", "SCHEMA"]),
-            ));
-        }
-    }
-
-    // Suggest find --task todo if there are open tasks.
-    let tasks_total = data
-        .get("tasks")
-        .and_then(|t| t.get("total"))
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let tasks_done = data
-        .get("tasks")
-        .and_then(|t| t.get("done"))
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if tasks_total > tasks_done {
-        hints.push(Hint::new(
-            "Find files with open tasks",
-            build_command_with_glob(ctx, &["find", "--task", "todo"]),
-        ));
-    }
-
-    // Suggest find --orphan if there are orphan files.
-    let orphan_count = data
-        .get("orphans")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if orphan_count > 0 && hints.len() < MAX_HINTS {
-        hints.push(Hint::new(
-            format!("{orphan_count} orphan files"),
-            build_command_with_glob(ctx, &["find", "--orphan"]),
-        ));
-    }
-
-    // Suggest find --dead-end if there are dead-end files.
-    let dead_end_count = data
-        .get("dead_ends")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if dead_end_count > 0 && hints.len() < MAX_HINTS {
-        hints.push(Hint::new(
-            format!("{dead_end_count} dead-end files"),
-            build_command_with_glob(ctx, &["find", "--dead-end"]),
-        ));
-    }
-
-    // Suggest find --broken-links if there are broken links.
-    let broken_links = data
-        .get("links")
-        .and_then(|l| l.get("broken"))
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let total_links = data
-        .get("links")
-        .and_then(|l| l.get("total"))
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    // Site-URL heuristic: when nearly every link is "broken" on a link-heavy
-    // vault, the links are almost certainly absolute site URLs the vault has no
-    // `--site-prefix` configured to resolve (MDN: 49,933/49,935 "broken"), not
-    // fixable file typos. Offering `links fix` on 50k links is actively
-    // misleading, so emit a diagnostic and skip the fix suggestion.
-    // Integer form of `broken/total >= SITE_URL_BROKEN_PERCENT/100` — avoids a
-    // lossy u64→f64 cast. `saturating_mul` guards the (implausible) overflow at
-    // >1.8e17 links.
-    let looks_like_site_urls = broken_links >= SITE_URL_MIN_BROKEN
-        && total_links > 0
-        && broken_links.saturating_mul(100) >= total_links.saturating_mul(SITE_URL_BROKEN_PERCENT);
-    if broken_links > 0 && hints.len() < MAX_HINTS {
-        hints.push(Hint::new(
-            format!("{broken_links} broken links"),
-            build_command_with_glob(ctx, &["find", "--broken-links"]),
-        ));
-        if looks_like_site_urls {
-            if hints.len() < MAX_HINTS {
-                hints.push(Hint::without_cmd(format!(
-                    "{broken_links} of {total_links} links are unresolvable — they look like \
-                     absolute site URLs; set `--site-prefix` (or `[site] prefix` in .hyalo.toml) \
-                     to resolve them rather than running `links fix`"
-                )));
-            }
-        } else if hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                "Auto-fix broken links (dry run)",
-                build_command_with_glob(ctx, &["links", "fix"]),
-            ));
-        }
-    }
-
-    // When schema is defined but no violations, or when there's still room,
-    // add the general lint / types hints.
-    if let Some(schema_obj) = data.get("schema") {
-        let errors = schema_obj
-            .get("errors")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        let warnings = schema_obj
-            .get("warnings")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        if errors == 0 && warnings == 0 && hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                "Validate frontmatter against schema",
-                build_command_with_glob(ctx, &["lint"]),
-            ));
-        }
-        if hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                "Manage type schemas",
-                build_command_no_glob(ctx, &["types", "list"]),
-            ));
-        }
-    }
-
-    // Pick 1-2 most interesting status values.
-    if let Some(status_arr) = data.get("status").and_then(|s| s.as_array()) {
-        let mut groups: Vec<(&str, u8)> = status_arr
-            .iter()
-            .filter_map(|g| {
-                let value = g.get("value").and_then(|v| v.as_str())?;
-                Some((value, status_priority(value)))
-            })
-            .collect();
-        groups.sort_by_key(|&(_, p)| p);
-
-        let remaining = MAX_HINTS.saturating_sub(hints.len());
-        for (value, _) in groups.into_iter().take(remaining.min(2)) {
-            let filter = format!("status={value}");
-            hints.push(Hint::new(
-                format!("Filter by status: {value}"),
-                build_command_no_glob(ctx, &["find", "--property", &filter]),
-            ));
-        }
-    }
-
-    // Prepend the create-index hint (computed at the top of this function) so
-    // it is visible even when all MAX_HINTS slots are consumed by health hints
-    // (orphans / broken-links / links-fix) on large vaults (NEW-1). Inserting
-    // at position 0 displaces the last (lowest-priority) hint when the cap is
-    // already reached — the index hint has the highest user-visible payoff.
-    if let Some(ci_hint) = create_index_hint {
-        hints.insert(0, ci_hint);
-        if hints.len() > MAX_HINTS {
-            hints.pop();
-        }
-    }
-
-    hints
-}
-
-fn hints_for_properties_summary(
-    ctx: &HintContext,
-    data: &serde_json::Value,
-    total: Option<u64>,
-) -> Vec<Hint> {
-    let Some(arr) = data.as_array() else {
-        return vec![];
-    };
-
-    let mut hints = Vec::new();
-
-    // When output was truncated by the default limit (not an explicit --limit), suggest
-    // showing all results.
-    if !ctx.has_limit {
-        let shown = arr.len() as u64;
-        if let Some(t) = total
-            && shown < t
-        {
-            hints.push(Hint::new(
-                format!("Show all {t} properties (no limit)"),
-                build_command_with_glob(ctx, &["properties", "summary", "--limit", "0"]),
-            ));
-        }
-    }
-
-    // Sort by count descending, take top 3.
-    let mut entries: Vec<(&str, u64)> = arr
-        .iter()
-        .filter_map(|e| {
-            let name = e.get("name").and_then(|n| n.as_str())?;
-            let count = e
-                .get("count")
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(0);
-            Some((name, count))
-        })
-        .collect();
-    entries.sort_by_key(|e| std::cmp::Reverse(e.1));
-
-    for (name, count) in entries.into_iter().take(3) {
-        if hints.len() >= MAX_HINTS {
-            break;
-        }
-        hints.push(Hint::new(
-            format!("Find {count} files with property: {name}"),
-            build_command_with_glob(ctx, &["find", "--property", name]),
-        ));
-    }
-
-    hints
-}
-
-/// Slugify a string to the charset valid for view names: `[a-z0-9_-]`.
-/// Replaces invalid chars with `-`, collapses runs of `-`, and trims leading/trailing `-`.
-fn slugify(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    for ch in s.chars() {
-        if ch.is_ascii_alphanumeric() || ch == '_' {
-            out.push(ch.to_ascii_lowercase());
-        } else {
-            // Replace any non-allowed char with a hyphen (collapsed below).
-            if !out.ends_with('-') {
-                out.push('-');
-            }
-        }
-    }
-    out.trim_matches('-').to_owned()
-}
-
-/// Format a confidence floor (0.0-1.0) as a `--min-confidence` CLI argument.
-///
-/// Rounds to 3 decimal places and trims trailing zeros/the decimal point, so
-/// `0.8` stays `"0.8"` rather than growing float-repr noise like
-/// `"0.8000000000000001"`, and `0.5` doesn't print as `"0.500"`.
-fn format_confidence(v: f64) -> String {
-    let mut s = format!("{v:.3}");
-    if s.contains('.') {
-        while s.ends_with('0') {
-            s.pop();
-        }
-        if s.ends_with('.') {
-            s.pop();
-        }
-    }
-    s
-}
-
-/// Derive a short, human-readable name from the active filters.
-fn auto_view_name(ctx: &HintContext) -> String {
-    let mut parts: Vec<String> = Vec::new();
-
-    for pf in &ctx.property_filters {
-        if let Some(pos) = pf.find("~=") {
-            // Regex filter (K~=pattern): use the key, not the pattern.
-            let key = &pf[..pos];
-            parts.push(key.to_lowercase());
-        } else if let Some(pos) = pf.find('=') {
-            let val = &pf[pos + 1..];
-            if !val.is_empty() {
-                parts.push(val.to_lowercase());
-            }
-        } else if let Some(stripped) = pf.strip_prefix('!') {
-            parts.push(format!("no-{stripped}"));
-        }
-    }
-
-    for tf in &ctx.tag_filters {
-        parts.push(tf.to_lowercase());
-    }
-
-    if let Some(task) = &ctx.task_filter {
-        parts.push(task.to_lowercase());
-    }
-
-    let slug = slugify(&parts.join("-"));
-    let truncated: String = slug.chars().take(40).collect();
-    // Trim any trailing `-` left by truncation mid-word.
-    let trimmed = truncated.trim_end_matches('-');
-    if trimmed.is_empty() {
-        "my-view".to_owned()
-    } else {
-        trimmed.to_owned()
-    }
-}
-
-/// Build the `hyalo views set <name> <filters…>` command string.
-fn build_views_set_command(ctx: &HintContext, view_name: &str) -> String {
-    let mut b = HintBuilder::cmd("views set");
-    b.push_quoted(view_name);
-    for pf in &ctx.property_filters {
-        b.push_raw("--property");
-        b.push_quoted(pf);
-    }
-    for tf in &ctx.tag_filters {
-        b.push_raw("--tag");
-        b.push_quoted(tf);
-    }
-    if let Some(task) = &ctx.task_filter {
-        b.push_raw("--task");
-        b.push_quoted(task);
-    }
-    b.finish(ctx)
-}
-
-/// Suggest saving the current query as a view when at least two
-/// view-serializable filter dimensions are active and the query did not
-/// itself come from a view. Excludes body/regex search since the actual
-/// pattern value is not available in `HintContext`.
-fn suggest_save_as_view(ctx: &HintContext) -> Option<Hint> {
-    if ctx.view_name.is_some() {
-        return None;
-    }
-
-    // Only count filters that can be round-tripped into a `views set` command.
-    // Body/regex search is excluded because `views set` does not support them,
-    // not the actual pattern string.
-    let filter_count =
-        ctx.property_filters.len() + ctx.tag_filters.len() + usize::from(ctx.task_filter.is_some());
-
-    if filter_count < 2 {
-        return None;
-    }
-
-    let name = auto_view_name(ctx);
-    let cmd = build_views_set_command(ctx, &name);
-    Some(Hint::new("Save this query as a view", cmd))
-}
-
-fn hints_for_find(ctx: &HintContext, data: &serde_json::Value, total: Option<u64>) -> Vec<Hint> {
-    // find returns a bare array as the raw command output (the envelope is built later).
-    let Some(results) = data.as_array() else {
-        return vec![];
-    };
-
-    if results.is_empty() {
-        // When a multi-word BM25 search returns nothing, suggest trying OR instead.
-        // Skip if the query already contains quotes (phrase search) — splitting on
-        // whitespace would produce malformed tokens like `"exact` and `phrase"`.
-        if let Some(pat) = &ctx.body_pattern {
-            let has_quotes = pat.contains('"');
-            let words: Vec<&str> = pat
-                .split_whitespace()
-                .filter(|w| {
-                    !w.starts_with('-')
-                        && !w.eq_ignore_ascii_case("or")
-                        && !w.eq_ignore_ascii_case("and")
-                })
-                .collect();
-            if !has_quotes && words.len() >= 2 {
-                let or_query = words.join(" OR ");
-                return vec![Hint::new(
-                    "Try OR instead of AND (match any word)",
-                    build_find_command_with_pattern(ctx, &or_query),
-                )];
-            }
-        }
-        return vec![];
-    }
-
-    let mut hints = Vec::new();
-    let result_count = results.len();
-    let is_single = result_count == 1;
-
-    // --- Single-result hints ---
-    if let Some(first_file) = results[0].get("file").and_then(|f| f.as_str()) {
-        hints.push(Hint::new(
-            "Read this file's content",
-            build_command_with_file(ctx, &["read"], first_file, &[]),
-        ));
-        if is_single {
-            hints.push(Hint::new(
-                "See all metadata for this file",
-                build_command_no_glob(ctx, &["find", "--file", first_file, "--fields", "all"]),
-            ));
-        }
-        hints.push(Hint::new(
-            "See what links to this file",
-            build_command_with_file(ctx, &["backlinks"], first_file, &[]),
-        ));
-    }
-
-    // --- Task bulk operation hints ---
-    // When find results target a single file and include task data, suggest bulk task ops.
-    if ctx.file_targets.len() == 1 {
-        let file = &ctx.file_targets[0];
-        let has_open_tasks = results.iter().any(|item| {
-            item.get("tasks")
-                .and_then(|t| t.as_array())
-                .is_some_and(|tasks| {
-                    tasks
-                        .iter()
-                        .any(|t| t.get("done") == Some(&serde_json::Value::Bool(false)))
-                })
-        });
-        if has_open_tasks {
-            let remaining = MAX_HINTS.saturating_sub(hints.len());
-            if remaining > 0 {
-                if let Some(section) = ctx.section_filters.first() {
-                    hints.push(Hint::new(
-                        format!("Toggle all tasks in section \"{section}\""),
-                        build_command_with_file(
-                            ctx,
-                            &["task", "toggle"],
-                            file,
-                            &["--section", section],
-                        ),
-                    ));
-                } else {
-                    hints.push(Hint::new(
-                        "Toggle all tasks in this file",
-                        build_command_with_file(ctx, &["task", "toggle"], file, &["--all"]),
-                    ));
-                }
-            }
-        }
-    }
-
-    // --- Broad query → suggest summary ---
-    let has_no_filters = ctx.property_filters.is_empty()
-        && ctx.tag_filters.is_empty()
-        && ctx.task_filter.is_none()
-        && !ctx.has_body_search
-        && !ctx.has_regex_search
-        && ctx.file_targets.is_empty();
-
-    if has_no_filters && result_count > 10 {
-        hints.push(Hint::new(
-            if ctx.glob.is_empty() {
-                "Get a high-level vault overview"
-            } else {
-                "Get stats for this file set"
-            },
-            build_command_with_glob(ctx, &["summary"]),
-        ));
-    }
-
-    // --- Show-all hint when default limit truncated output ---
-    if !ctx.has_limit
-        && let Some(t) = total
-        && (result_count as u64) < t
-    {
-        let remaining = MAX_HINTS.saturating_sub(hints.len());
-        if remaining > 0 {
-            hints.push(Hint::new(
-                format!("Show all {t} results (no limit)"),
-                build_find_command_preserving_filters(ctx, &["--limit", "0"]),
-            ));
-        }
-    }
-
-    // --- Narrowing for many results (>5) ---
-    if result_count > 5 {
-        // Tag narrowing (skip tags already filtered on).
-        let mut tag_counts: std::collections::HashMap<&str, usize> =
-            std::collections::HashMap::new();
-        for item in results {
-            if let Some(tags) = item.get("tags").and_then(|t| t.as_array()) {
-                for tag in tags {
-                    if let Some(name) = tag.as_str()
-                        && !ctx.tag_filters.iter().any(|t| t == name)
-                    {
-                        *tag_counts.entry(name).or_insert(0) += 1;
-                    }
-                }
-            }
-        }
-
-        // Collect status property frequencies — skip statuses already filtered on.
-        // Handles both scalar and array-valued status properties.
-        let mut status_counts: std::collections::HashMap<&str, usize> =
-            std::collections::HashMap::new();
-        for item in results {
-            let Some(status_val) = item.get("properties").and_then(|p| p.get("status")) else {
-                continue;
-            };
-            // Yield individual &str values from scalar or array status.
-            let iter: Box<dyn Iterator<Item = &str>> = match status_val {
-                serde_json::Value::String(s) => Box::new(std::iter::once(s.as_str())),
-                serde_json::Value::Array(arr) => Box::new(arr.iter().filter_map(|v| v.as_str())),
-                _ => Box::new(std::iter::empty()),
-            };
-            for status in iter {
-                let already_filtered = ctx
-                    .property_filters
-                    .iter()
-                    .any(|f| f == &format!("status={status}"));
-                if !already_filtered {
-                    *status_counts.entry(status).or_insert(0) += 1;
-                }
-            }
-        }
-
-        // Whether the shown results are only a page of a larger set. When they
-        // are, the per-tag / per-status frequencies below were computed on the
-        // truncated page, so they undercount the true filtered totals — we must
-        // not present them as the count the composed command would return
-        // (BUG-8: hint said 27, the command returned 37). In that case the
-        // parenthetical count is dropped and the hint carries the composing
-        // command only.
-        let is_truncated = total.is_some_and(|t| (result_count as u64) < t);
-
-        // Pick the most common tag (if any results have tags).
-        // Break ties alphabetically for deterministic output.
-        if let Some((top_tag, count)) = tag_counts
-            .iter()
-            .max_by(|(a_tag, a_cnt), (b_tag, b_cnt)| a_cnt.cmp(b_cnt).then(b_tag.cmp(a_tag)))
-        {
-            let remaining = MAX_HINTS.saturating_sub(hints.len());
-            if remaining > 0 {
-                // Compose with the active filters (BUG-8): a "narrow by tag"
-                // hint on a `--orphan` query must keep `--orphan`, else pasting
-                // it widens the set.
-                let description = if is_truncated {
-                    format!("Narrow by tag: {top_tag}")
-                } else {
-                    format!("Narrow by tag: {top_tag} ({count} files)")
-                };
-                hints.push(Hint::new(
-                    description,
-                    build_find_command_composing(ctx, &["--tag", top_tag]),
-                ));
-            }
-        }
-
-        // Pick the most interesting status value (prefer active/planned over completed).
-        let mut status_vec: Vec<(&str, usize, u8)> = status_counts
-            .iter()
-            .map(|(v, c)| (*v, *c, status_priority(v)))
-            .collect();
-        // Sort by priority (ascending), then count (descending), then name (ascending).
-        status_vec.sort_by(|a, b| a.2.cmp(&b.2).then(b.1.cmp(&a.1)).then(a.0.cmp(b.0)));
-
-        if let Some((top_status, count, _)) = status_vec.first() {
-            let remaining = MAX_HINTS.saturating_sub(hints.len());
-            if remaining > 0 {
-                let status_filter = format!("status={top_status}");
-                let description = if is_truncated {
-                    format!("Filter by status: {top_status}")
-                } else {
-                    format!("Filter by status: {top_status} ({count} files)")
-                };
-                hints.push(Hint::new(
-                    description,
-                    build_find_command_composing(ctx, &["--property", &status_filter]),
-                ));
-            }
-        }
-
-        // Sort suggestion (only if not already sorting).
-        if ctx.sort.is_none() {
-            let remaining = MAX_HINTS.saturating_sub(hints.len());
-            if remaining > 0 {
-                hints.push(Hint::new(
-                    "Sort by most recently modified",
-                    build_find_command_preserving_filters(
-                        ctx,
-                        &["--sort", "modified", "--reverse"],
-                    ),
-                ));
-            }
-        }
-
-        // Limit suggestion: suggest --limit 10 when not truncated and no explicit limit.
-        if !ctx.has_limit && total.is_none_or(|t| (result_count as u64) >= t) {
-            let remaining = MAX_HINTS.saturating_sub(hints.len());
-            if remaining > 0 {
-                hints.push(Hint::new(
-                    "Limit to 10 results",
-                    build_find_command_preserving_filters(ctx, &["--limit", "10"]),
-                ));
-            }
-        }
-    }
-
-    // Suggest saving as a view for non-trivial queries (independent of result count).
-    if let Some(view_hint) = suggest_save_as_view(ctx) {
-        let remaining = MAX_HINTS.saturating_sub(hints.len());
-        if remaining > 0 {
-            hints.push(view_hint);
-        }
-    }
-
-    // Body search → regex suggestion is intentionally omitted.
-    // We cannot produce a concrete regex without knowing the user's intent,
-    // and a placeholder like `'pattern'` would violate our no-templates contract.
-
-    // Suggest phrase search when body search has multiple words and many results.
-    if let Some(pat) = &ctx.body_pattern {
-        let has_quotes = pat.contains('"');
-        let words: Vec<&str> = pat
-            .split_whitespace()
-            .filter(|w| {
-                !w.starts_with('-')
-                    && !w.eq_ignore_ascii_case("or")
-                    && !w.eq_ignore_ascii_case("and")
-            })
-            .collect();
-        if !has_quotes && words.len() >= 2 && result_count > 10 {
-            let remaining = MAX_HINTS.saturating_sub(hints.len());
-            if remaining > 0 {
-                let phrase = format!("\"{}\"", words.join(" "));
-                hints.push(Hint::new(
-                    "Try as exact phrase for more precise results",
-                    build_find_command_with_pattern(ctx, &phrase),
-                ));
-            }
-        }
-    }
-
-    // Suggest `links fix` when results contain broken links (e.g. from --broken-links).
-    // Broken links are serialised with `"path": null` (never omitted) by find's output.
-    let has_broken_links = results.iter().any(|item| {
-        item.get("links")
-            .and_then(|l| l.as_array())
-            .is_some_and(|links| {
-                links
-                    .iter()
-                    .any(|link| link.get("path").is_some_and(serde_json::Value::is_null))
-            })
-    });
-    if has_broken_links {
-        let remaining = MAX_HINTS.saturating_sub(hints.len());
-        if remaining > 0 {
-            hints.push(Hint::new(
-                "Auto-fix broken links (dry run)",
-                build_command_with_glob(ctx, &["links", "fix"]),
-            ));
-        }
-    }
-
-    hints
-}
-
-fn hints_for_tags_summary(
-    ctx: &HintContext,
-    data: &serde_json::Value,
-    total: Option<u64>,
-) -> Vec<Hint> {
-    // tags summary returns a bare array [{name, count}, ...].
-    let Some(tags_arr) = data.as_array() else {
-        return vec![];
-    };
-
-    let mut hints = Vec::new();
-
-    // When output was truncated by the default limit (not an explicit --limit), suggest
-    // showing all results.
-    if !ctx.has_limit {
-        let shown = tags_arr.len() as u64;
-        if let Some(t) = total
-            && shown < t
-        {
-            hints.push(Hint::new(
-                format!("Show all {t} tags (no limit)"),
-                build_command_with_glob(ctx, &["tags", "summary", "--limit", "0"]),
-            ));
-        }
-    }
-
-    // Sort by count descending, take top 3.
-    let mut entries: Vec<(&str, u64)> = tags_arr
-        .iter()
-        .filter_map(|entry| {
-            let name = entry.get("name").and_then(|n| n.as_str())?;
-            let count = entry
-                .get("count")
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(0);
-            Some((name, count))
-        })
-        .collect();
-    entries.sort_by_key(|e| std::cmp::Reverse(e.1));
-
-    for (name, count) in entries.into_iter().take(3) {
-        if hints.len() >= MAX_HINTS {
-            break;
-        }
-        hints.push(Hint::new(
-            format!("Find {count} files tagged: {name}"),
-            build_command_with_glob(ctx, &["find", "--tag", name]),
-        ));
-    }
-
-    hints
-}
-
-fn hints_for_mutation(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let first_modified = first_modified_file(data);
-
-    if let Some(file) = first_modified {
-        hints.push(Hint::new(
-            "Verify the updated file",
-            build_command_no_glob(
-                ctx,
-                &["find", "--file", file, "--fields", "properties,tags"],
-            ),
-        ));
-        hints.push(Hint::new(
-            "Read the modified file",
-            build_command_no_glob(ctx, &["read", file]),
-        ));
-    }
-
-    hints
-}
-
-fn hints_for_read(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let file = data
-        .get("file")
-        .and_then(|f| f.as_str())
-        .or_else(|| ctx.file_targets.first().map(String::as_str));
-
-    if let Some(file) = file {
-        hints.push(Hint::new(
-            "See metadata for this file",
-            build_command_no_glob(ctx, &["find", "--file", file, "--fields", "all"]),
-        ));
-        hints.push(Hint::new(
-            "See what links to this file",
-            build_command_with_file(ctx, &["backlinks"], file, &[]),
-        ));
-        // UX-4 (dogfood pre3): `read --format json` without `--frontmatter`
-        // silently omits the `frontmatter` key entirely — `--jq
-        // '.results.frontmatter.x'` reads that as `null` indistinguishably
-        // from "the property doesn't exist," costing a round trip to
-        // discover the flag was the actual gap. Only worth saying when the
-        // key really is missing (a caller who already asked for it needs no
-        // reminder).
-        if data.get("frontmatter").is_none() && hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                "Include frontmatter in the output",
-                build_command_with_file(ctx, &["read"], file, &["--frontmatter"]),
-            ));
-        }
-    }
-
-    hints
-}
-
-fn hints_for_backlinks(
-    ctx: &HintContext,
-    data: &serde_json::Value,
-    total: Option<u64>,
-) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    // When output was truncated by the default limit (not an explicit --limit), suggest
-    // showing all results.
-    if !ctx.has_limit {
-        let shown = data
-            .get("backlinks")
-            .and_then(|b| b.as_array())
-            .map_or(0, |a| a.len() as u64);
-        if let Some(t) = total
-            && shown < t
-        {
-            let file = data.get("file").and_then(|f| f.as_str()).unwrap_or("");
-            hints.push(Hint::new(
-                format!("Show all {t} backlinks (no limit)"),
-                build_command_with_file(ctx, &["backlinks", "--limit", "0"], file, &[]),
-            ));
-        }
-    }
-
-    let file = data.get("file").and_then(|f| f.as_str());
-
-    if let Some(file) = file {
-        hints.push(Hint::new(
-            "Read this file's content",
-            build_command_with_file(ctx, &["read"], file, &[]),
-        ));
-        hints.push(Hint::new(
-            "See this file's outgoing links",
-            build_command_no_glob(ctx, &["find", "--file", file, "--fields", "links"]),
-        ));
-    }
-
-    // Suggest reading the first backlink source.
-    if let Some(backlinks) = data.get("backlinks").and_then(|b| b.as_array())
-        && let Some(first_source) = backlinks
-            .first()
-            .and_then(|b| b.get("source"))
-            .and_then(|s| s.as_str())
-        && hints.len() < MAX_HINTS
-    {
-        hints.push(Hint::new(
-            format!("Read linking file: {first_source}"),
-            build_command_with_file(ctx, &["read"], first_source, &[]),
-        ));
-    }
-
-    hints
-}
-
-fn hints_for_mv(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let to_path = data.get("to").and_then(|t| t.as_str());
-    let is_dry_run = data
-        .get("dry_run")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-
-    if let Some(to_path) = to_path {
-        if is_dry_run {
-            if let Some(from_path) = data.get("from").and_then(|f| f.as_str()) {
-                hints.push(Hint::new(
-                    "Apply this move",
-                    build_command_with_file(ctx, &["mv"], from_path, &["--to", to_path]),
-                ));
-            }
-        } else {
-            hints.push(Hint::new(
-                "Read the moved file",
-                build_command_with_file(ctx, &["read"], to_path, &[]),
-            ));
-            hints.push(Hint::new(
-                "Verify backlinks updated",
-                build_command_with_file(ctx, &["backlinks"], to_path, &[]),
-            ));
-        }
-    }
-
-    hints
-}
-
-/// Check if task output data (single or array) contains any open (not done) tasks.
-fn task_result_has_open(data: &serde_json::Value) -> bool {
-    // Array case (bulk result)
-    if let Some(arr) = data.as_array() {
-        return arr
-            .iter()
-            .any(|t| t.get("done") == Some(&serde_json::Value::Bool(false)));
-    }
-    // Single task case
-    data.get("done") == Some(&serde_json::Value::Bool(false))
-}
-
-/// Hints for `task read` — suggest toggling or viewing remaining tasks.
-fn hints_for_task_read(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    // For bulk reads (--all / --section), suggest toggling the same scope.
-    if let Some(selector) = &ctx.task_selector {
-        if let Some(file) = ctx.file_targets.first() {
-            let has_open = task_result_has_open(data);
-            if has_open {
-                if selector == "all" {
-                    hints.push(Hint::new(
-                        "Toggle all tasks in this file",
-                        build_command_with_file(ctx, &["task", "toggle"], file, &["--all"]),
-                    ));
-                } else if let Some(section) = selector.strip_prefix("section:") {
-                    hints.push(Hint::new(
-                        format!("Toggle all tasks in section \"{section}\""),
-                        build_command_with_file(
-                            ctx,
-                            &["task", "toggle"],
-                            file,
-                            &["--section", section],
-                        ),
-                    ));
-                }
-            } else {
-                // NEW-18 (dogfood pre3): a bulk read (`--all` / `--section`)
-                // whose tasks are all already done had nothing to toggle and
-                // fell straight through to an empty hint list — a listing
-                // command that answered "nothing here" was still a
-                // navigation dead end instead of pointing anywhere else.
-                hints.push(Hint::new(
-                    "Find files with open tasks",
-                    build_command_no_glob(ctx, &["find", "--task", "todo"]),
-                ));
-            }
-        }
-        // For "all" and "section:" selectors, return early — the bulk hints are sufficient.
-        // For "lines" selector, fall through to the single-task hint path which handles
-        // individual line-based suggestions.
-        if selector != "lines" {
-            return hints;
-        }
-    }
-
-    // Single-task read path (backward compatible).
-    let file = data.get("file").and_then(|f| f.as_str());
-    let line = data.get("line").and_then(serde_json::Value::as_u64);
-    let done = data
-        .get("done")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-
-    if let (Some(file), Some(line)) = (file, line) {
-        let line_str = line.to_string();
-        if !done {
-            hints.push(Hint::new(
-                "Toggle this task to done",
-                build_command_with_file(ctx, &["task", "toggle"], file, &["--line", &line_str]),
-            ));
-        }
-        hints.push(Hint::new(
-            "See all open tasks in this file",
-            build_command_no_glob(
-                ctx,
-                &[
-                    "find", "--file", file, "--task", "todo", "--fields", "tasks",
-                ],
-            ),
-        ));
-    }
-
-    hints
-}
-
-fn hints_for_task_mutation(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let file = ctx
-        .file_targets
-        .first()
-        .map(String::as_str)
-        .or_else(|| data.get("file").and_then(|f| f.as_str()));
-
-    if let Some(file) = file {
-        // Suggest reading the scope that was just mutated.
-        if let Some(selector) = &ctx.task_selector {
-            if selector == "all" {
-                hints.push(Hint::new(
-                    "Read all tasks in this file",
-                    build_command_with_file(ctx, &["task", "read"], file, &["--all"]),
-                ));
-            } else if let Some(section) = selector.strip_prefix("section:") {
-                hints.push(Hint::new(
-                    format!("Read tasks in section \"{section}\""),
-                    build_command_with_file(ctx, &["task", "read"], file, &["--section", section]),
-                ));
-            }
-        }
-
-        hints.push(Hint::new(
-            "See remaining open tasks",
-            build_command_no_glob(
-                ctx,
-                &[
-                    "find", "--file", file, "--task", "todo", "--fields", "tasks",
-                ],
-            ),
-        ));
-        hints.push(Hint::new(
-            "Read the file",
-            build_command_with_file(ctx, &["read"], file, &[]),
-        ));
-    }
-
-    hints
-}
-
-fn hints_for_links_fix(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let is_dry_run = !data
-        .get("applied")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    let fixable = data
-        .get("fixable")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let unfixable = data
-        .get("unfixable")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    // L-25: dry-run validates plans against on-disk text, so some `fixable`
-    // fixes may be stale (their text no longer matches). Discount them so the
-    // "Apply N fixes" hint count matches what `--apply` actually writes.
-    let unapplied = data
-        .get("unapplied")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let applicable = fixable.saturating_sub(unapplied);
-
-    if is_dry_run && applicable > 0 {
-        hints.push(Hint::new(
-            format!("Apply {applicable} fixes"),
-            build_command_with_glob(ctx, &["links", "fix", "--apply"]),
-        ));
-    }
-
-    // Fuzzy-match fixes are excluded from --apply by default. Surface the
-    // opt-in when a dry-run turned up fuzzy candidates that were not applied
-    // — but only count candidates that actually clear the confidence floor
-    // (NEW-14, iter-218). `fuzzy` is every candidate found, including ones
-    // below the floor that `--apply-fuzzy` would never write; promising to
-    // apply all of `fuzzy` produced a hint like "apply 3253 fixes" that
-    // applied 0 files when every one of them was below-floor.
-    let fuzzy = data
-        .get("fuzzy")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let fuzzy_below_floor = data
-        .get("fuzzy_below_floor")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let applicable_fuzzy = fuzzy.saturating_sub(fuzzy_below_floor);
-    let fuzzy_applied = data
-        .get("fuzzy_applied")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    if applicable_fuzzy > 0 && !fuzzy_applied {
-        // `applicable_fuzzy` is counted against THIS run's effective floor
-        // (`fuzzy_min_confidence`), which may differ from
-        // `DEFAULT_FUZZY_MIN_CONFIDENCE` — either because this run passed
-        // `--min-confidence` or because `.hyalo.toml` sets
-        // `[links] fuzzy_min_confidence`. The hinted command must carry the
-        // same floor, or a dry run at a lower floor promises a count the
-        // hinted apply (which would silently fall back to the default 0.8)
-        // does not deliver (review finding #1). Appending it whenever the
-        // floor is non-default is simpler than tracking whether it came
-        // from the flag or the config file, and is harmless when it came
-        // from the config (the flag just repeats what the config would
-        // have applied anyway).
-        let floor = data
-            .get("fuzzy_min_confidence")
-            .and_then(serde_json::Value::as_f64)
-            .unwrap_or(hyalo_core::link_score::DEFAULT_FUZZY_MIN_CONFIDENCE);
-        let mut cmd_parts = vec!["links", "fix", "--apply", "--apply-fuzzy"];
-        let floor_arg;
-        if (floor - hyalo_core::link_score::DEFAULT_FUZZY_MIN_CONFIDENCE).abs() > f64::EPSILON {
-            floor_arg = format_confidence(floor);
-            cmd_parts.push("--min-confidence");
-            cmd_parts.push(&floor_arg);
-        }
-        hints.push(Hint::new(
-            format!("Review then apply {applicable_fuzzy} lower-confidence fuzzy fixes"),
-            build_command_with_glob(ctx, &cmd_parts),
-        ));
-    } else if fuzzy > 0 && !fuzzy_applied {
-        // Every candidate is below the floor — `--apply-fuzzy` would apply 0
-        // of them. Point at reviewing with a lower floor instead of a
-        // command that reads as an apply but writes nothing.
-        let floor = data
-            .get("fuzzy_min_confidence")
-            .and_then(serde_json::Value::as_f64)
-            .unwrap_or(0.0);
-        hints.push(Hint::new(
-            format!(
-                "{fuzzy} fuzzy candidates found, all below the confidence floor \
-                 {floor} — review with a lower --min-confidence before applying"
-            ),
-            build_command_with_glob(ctx, &["links", "fix", "--min-confidence", "0"]),
-        ));
-    }
-
-    if unfixable > 0 {
-        hints.push(Hint::new(
-            "List files with remaining broken links",
-            build_command_with_glob(ctx, &["find", "--broken-links"]),
-        ));
-    }
-
-    // L-11: a non-zero `failed` count means some fixes produced a valid plan
-    // but the durable write itself failed mid-batch (e.g. read-only file).
-    // Point at the per-fix detail and suggest the most common cause.
-    let failed = data
-        .get("failed")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if failed > 0 {
-        hints.push(Hint::without_cmd(format!(
-            "{failed} fix(es) failed to write — see `failed_fixes` for the per-file \
-             error, check file permissions"
-        )));
-    }
-
-    // Case-mismatch and relocation repairs are written by plain `--apply` but
-    // are *not* part of `fixable`, so a vault whose only problem was one of
-    // these produced no "Apply" hint at all — the fix was available and
-    // unadvertised (iter-210). NEW-13 (dogfood pre3) split relocations out of
-    // `case_mismatches` into their own bucket; both still land here since
-    // both are written by plain `--apply`.
-    let case_mismatches = data
-        .get("case_mismatches")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let relocations = data
-        .get("relocations")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    let case_and_relocation_fixes = case_mismatches + relocations;
-    if is_dry_run && case_and_relocation_fixes > 0 && applicable == 0 && hints.len() < MAX_HINTS {
-        let label = if case_mismatches > 0 && relocations > 0 {
-            format!("Apply {case_mismatches} case-mismatch and {relocations} relocation fixes")
-        } else if relocations > 0 {
-            format!("Apply {relocations} relocation fixes")
-        } else {
-            format!("Apply {case_mismatches} case-mismatch fixes")
-        };
-        hints.push(Hint::new(
-            label,
-            build_command_with_glob(ctx, &["links", "fix", "--apply"]),
-        ));
-    }
-
-    // A vault with nothing broken used to emit no hints whatsoever, making
-    // `links` a navigation dead end (dogfood UX-4). Point at the two link
-    // questions a clean fix report does *not* answer.
-    if hints.is_empty() {
-        hints.push(Hint::new(
-            "Preview title mentions that could become links",
-            build_command_with_glob(ctx, &["links", "auto"]),
-        ));
-        hints.push(Hint::new(
-            "List notes nothing links to",
-            build_command_no_glob(ctx, &["find", "--orphan"]),
-        ));
-    }
-
-    hints
-}
-
-fn hints_for_links_auto(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    // iter-216 D-4: prefer the explicit `dry_run` key. `applied` is false both
-    // for a preview and for an `--apply` run that found nothing, so inverting
-    // it alone mislabels the latter.
-    let is_dry_run = data
-        .get("dry_run")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or_else(|| {
-            !data
-                .get("applied")
-                .and_then(serde_json::Value::as_bool)
-                .unwrap_or(false)
-        });
-    // iter-216 D-3: `matched` (was `total`).
-    let total = data
-        .get("matched")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-
-    if is_dry_run && total > 0 {
-        // Rebuild the exact command from the preview, preserving all
-        // scope-narrowing flags so the apply doesn't widen the mutation set.
-        let mut args: Vec<&str> = vec!["links", "auto", "--apply"];
-        let min_str;
-        if let Some(ml) = ctx.auto_link_min_length
-            && ml != 3
-        {
-            args.push("--min-length");
-            min_str = ml.to_string();
-            args.push(&min_str);
-        }
-        let cmd = build_command_with_glob(ctx, &args);
-        // Append --file and --exclude-title after the builder (they are not
-        // glob-related and aren't handled by build_command_with_glob).
-        let mut parts = vec![cmd];
-        if let Some(ref f) = ctx.auto_link_file {
-            parts.push(format!("--file {}", shell_quote(f)));
-        }
-        for et in &ctx.auto_link_exclude_titles {
-            parts.push(format!("--exclude-title {}", shell_quote(et)));
-        }
-        hints.push(Hint::new(
-            format!("Apply {total} auto-links"),
-            parts.join(" "),
-        ));
-    }
-
-    // L-11: a non-zero `files_failed` count means some files produced a
-    // valid auto-link plan but the durable write itself failed mid-batch.
-    // Point at the per-file detail and suggest the most common cause.
-    let files_failed = data
-        .get("files_failed")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if files_failed > 0 {
-        hints.push(Hint::without_cmd(format!(
-            "{files_failed} file(s) failed to write — see `apply_outcomes` for the \
-             per-file error, check file permissions"
-        )));
-    }
-
-    hints
-}
-
-fn hints_for_create_index(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    // Use bare `--index` (defaults to .hyalo-index in vault dir) for the default path.
-    // Only include the explicit path when the index was created at a non-default location.
-    let index_path = data
-        .get("path")
-        .and_then(|p| p.as_str())
-        .or(ctx.index_path.as_deref());
-
-    // Only treat as default when no path was reported or it's the bare default name.
-    // Custom paths like `sub/.hyalo-index` must emit the explicit path in the hint.
-    let is_default = index_path.is_none_or(|p| p == ".hyalo-index");
-
-    let hint_cmd = if is_default {
-        build_command_no_glob(ctx, &["find", "--index"])
-    } else {
-        build_command_no_glob(
-            ctx,
-            &["find", "--index-file", index_path.unwrap_or(".hyalo-index")],
-        )
-    };
-
-    hints.push(Hint::new("Query using the index", hint_cmd));
-    // L-9: the drop hint has to name the same file the query hint does.
-    // A bare `drop-index` after `create-index -o <custom>` targets the default
-    // `<vault>/.hyalo-index` — a different file, which is either absent or
-    // someone else's index.
-    let drop_cmd = if is_default {
-        build_command_no_glob(ctx, &["drop-index"])
-    } else {
-        build_command_no_glob(
-            ctx,
-            &["drop-index", "--path", index_path.unwrap_or(".hyalo-index")],
-        )
-    };
-    hints.push(Hint::new("Delete the index when done", drop_cmd));
-
-    hints
-}
-
-fn hints_for_drop_index(ctx: &HintContext, _data: &serde_json::Value) -> Vec<Hint> {
-    vec![Hint::new(
-        "Rebuild the index",
-        build_command_no_glob(ctx, &["create-index"]),
-    )]
-}
-
-/// Build the profile-aware `hyalo lint` validate hint. When the `okf` profile is
-/// already active via `[lint] profiles`, the `--profile okf` flag is redundant
-/// (plain `hyalo lint` runs it), so it is dropped to avoid noisy advice.
-fn okf_validate_hint(ctx: &HintContext) -> Hint {
-    let cmd = if ctx.okf_profile_active {
-        build_command_no_glob(ctx, &["lint"])
-    } else {
-        build_command_no_glob(ctx, &["lint", "--profile", "okf"])
-    };
-    Hint::new("Validate bundle conformance", cmd)
-}
-
-/// Drill-down hints for `hyalo okf index`.
-///
-/// Always suggests validating conformance. In a dry run that detected drift, it
-/// additionally suggests applying the regenerated index. The `apply` /
-/// `changed` signals are read from the command's own JSON result so the hint
-/// stays in sync with what actually happened.
-fn hints_for_okf_index(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-    let applied = data
-        .get("apply")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    let changed = data
-        .get("changed")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    // On dry-run drift, applying is the natural next step — surface it first.
-    if !applied && changed > 0 {
-        hints.push(Hint::new(
-            "Apply the regenerated index files",
-            build_command_no_glob(ctx, &["okf", "index", "--apply"]),
-        ));
-    }
-    hints.push(okf_validate_hint(ctx));
-    hints
-}
-
-/// Drill-down hints for `hyalo okf log`: suggest validating conformance.
-fn hints_for_okf_log(ctx: &HintContext) -> Vec<Hint> {
-    vec![okf_validate_hint(ctx)]
-}
-
-/// Ratio threshold for rule dominance (UX-2).
-const RULE_DOMINANCE_RATIO: f64 = 0.5;
-/// Absolute minimum violations for rule dominance (UX-2).
-const RULE_DOMINANCE_MIN: usize = 50;
-
-/// Return a per-rule hint entry for HYALO001 or HYALO002, or `None` for other rules.
-fn per_rule_hint(ctx: &HintContext, rule_id: &str, worst_file: Option<&str>) -> Option<Hint> {
-    match rule_id {
-        "HYALO001" => Some(Hint::new(
-            "Auto-fix HYALO001 violations",
-            build_lint_with_filter_flags(ctx, &["lint", "--rule", "HYALO001", "--fix"]),
-        )),
-        "HYALO002" => worst_file.map(|file| {
-            // Use `--file <path>` rather than a positional, since `find`'s
-            // positional argument is the search pattern.
-            Hint::new(
-                format!("See open tasks in {file}"),
-                build_command_no_glob(ctx, &["find", "--task", "todo", "--file", file]),
-            )
-        }),
-        _ => None,
-    }
-}
-
-/// Build a lint command that preserves `--rule`, `--rule-prefix`, `--fix-rule`, glob, and
-/// file targets from the current context, then appends `args`.
-fn build_lint_with_filter_flags(ctx: &HintContext, args: &[&str]) -> String {
-    let mut b = HintBuilder::empty();
-    for arg in args {
-        b.push_quoted(arg);
-    }
-    // Preserve rule/prefix/fix-rule filters from the original invocation.
-    if let Some(rule) = &ctx.lint_rule
-        && !args.contains(&"--rule")
-    {
-        b.push_raw("--rule");
-        b.push_quoted(rule);
-    }
-    if let Some(prefix) = &ctx.lint_rule_prefix {
-        b.push_raw("--rule-prefix");
-        b.push_quoted(prefix);
-    }
-    for fr in &ctx.lint_fix_rules {
-        b.push_raw("--fix-rule");
-        b.push_quoted(fr);
-    }
-    for glob in &ctx.glob {
-        b.push_raw("--glob");
-        b.push_quoted(glob);
-    }
-    for ft in &ctx.file_targets {
-        b.push_quoted(ft);
-    }
-    b.finish(ctx)
-}
-
-/// Accumulate rule violation counts from a named array field in a file JSON object.
-fn accumulate_rule_groups(
-    file: &serde_json::Value,
-    key: &str,
-    totals: &mut std::collections::HashMap<String, usize>,
-) {
-    if let Some(groups) = file.get(key).and_then(|rg| rg.as_array()) {
-        for group in groups {
-            if let (Some(rule), Some(count)) = (
-                group.get("rule").and_then(serde_json::Value::as_str),
-                group.get("count").and_then(serde_json::Value::as_u64),
-            ) {
-                *totals.entry(rule.to_owned()).or_default() +=
-                    usize::try_from(count).unwrap_or(usize::MAX);
-            }
-        }
-    }
-}
-
-/// Collect per-rule violation counts across all files, scanning both `rule_groups` (read-only)
-/// and `fixed_groups` + `remaining_groups` (fix-mode).
-fn collect_rule_totals(data: &serde_json::Value) -> std::collections::HashMap<String, usize> {
-    let mut totals: std::collections::HashMap<String, usize> = std::collections::HashMap::new();
-    let Some(files) = data.get("files").and_then(|f| f.as_array()) else {
-        return totals;
-    };
-    for file in files {
-        // Read-only shape.
-        accumulate_rule_groups(file, "rule_groups", &mut totals);
-        // Fix-mode shapes (count all including fixed for dominance analysis).
-        accumulate_rule_groups(file, "remaining_groups", &mut totals);
-        accumulate_rule_groups(file, "fixed_groups", &mut totals);
-    }
-    totals
-}
-
-fn hints_for_lint(ctx: &HintContext, data: &serde_json::Value, _total: Option<u64>) -> Vec<Hint> {
-    let mut hints: Vec<Hint> = Vec::new();
-
-    let is_fix_mode = ctx.lint_is_fix;
-    let is_dry_run = ctx.dry_run
-        || data
-            .get("dry_run")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(false);
-
-    // -----------------------------------------------------------------------
-    // Show-all hint when output is truncated.
-    // -----------------------------------------------------------------------
-    // Suppressed after `--fix` apply: the `files_with_issues` count reflects
-    // the pre-fix state, but the hinted `hyalo lint --limit 0` drops `--fix`
-    // and re-lints the now-fixed vault, so the count no longer matches (BUG-9
-    // carryover / "Post-`lint --fix` output drops the stale hint"). Read-only
-    // and dry-run modes keep it because their counts are still accurate.
-    let is_apply = is_fix_mode && !is_dry_run;
-    let is_limited = data
-        .get("files_truncated")
-        .or_else(|| data.get("limited"))
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    if !ctx.has_limit && is_limited && !is_apply {
-        let total_violations = data
-            .get("files_with_violations")
-            .or_else(|| data.get("files_with_issues"))
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        // UX-4 (dogfood v0.20.0): when errors exist anywhere in the run but
-        // the truncated `files[]` slice shows none (or not all) of them,
-        // say so in the hint — a "4 errors" count in the summary line with
-        // zero visible error lines reads like a bug otherwise. With the
-        // errors-first display sort this only fires when there are more
-        // error-carrying files than the cap, but it keeps the listing honest
-        // in that case too.
-        // Review M-1 (PR #277): read-only mode only — fix-mode JSON shapes
-        // the per-file groups as `fixed_groups`/`remaining_groups` (no
-        // `rule_groups`), so this computation would wrongly report every
-        // error as hidden. Per-file error totals are derived from each
-        // group's `count` (the per-rule display cap truncates `violations[]`
-        // but never `count`), so the hidden figure reflects the *file* cap.
-        let listed_errors: u64 = if is_fix_mode {
-            u64::MAX // read-only-only suffix: never claims hidden errors in fix mode (total_errors > u64::MAX is never true)
-        } else {
-            data.get("files")
-                .and_then(|f| f.as_array())
-                .map_or(0, |files| {
-                    files
-                        .iter()
-                        .map(|file| {
-                            file.get("rule_groups").and_then(|rg| rg.as_array()).map_or(
-                                0,
-                                |groups| {
-                                    groups
-                                        .iter()
-                                        .filter(|g| {
-                                            g.get("severity").and_then(serde_json::Value::as_str)
-                                                == Some("error")
-                                        })
-                                        .map(|g| {
-                                            g.get("count")
-                                                .and_then(serde_json::Value::as_u64)
-                                                .unwrap_or(0)
-                                        })
-                                        .sum::<u64>()
-                                },
-                            )
-                        })
-                        .sum()
-                })
-        };
-        let total_errors = data
-            .get("errors")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        let description = if total_errors > listed_errors {
-            let hidden = total_errors - listed_errors;
-            let err_label = if hidden == 1 { "error" } else { "errors" };
-            format!(
-                "Show all {total_violations} files with issues ({hidden} {err_label} hidden by the file cap)"
-            )
-        } else {
-            format!("Show all {total_violations} files with issues (no limit)")
-        };
-        hints.push(Hint::new(
-            description,
-            build_command_with_glob_and_files(ctx, &["lint", "--limit", "0"]),
-        ));
-    }
-
-    // -----------------------------------------------------------------------
-    // UX-7: Smart fix/dry-run hints.
-    // -----------------------------------------------------------------------
-    if !is_fix_mode {
-        // Not in fix mode: suggest preview (don't suggest apply directly).
-        let has_violations = data
-            .get("files")
-            .and_then(|f| f.as_array())
-            .is_some_and(|files| {
-                files.iter().any(|file| {
-                    file.get("rule_groups")
-                        .and_then(|rg| rg.as_array())
-                        .is_some_and(|groups| {
-                            groups.iter().any(|g| {
-                                g.get("autofixable")
-                                    .and_then(serde_json::Value::as_bool)
-                                    .unwrap_or(false)
-                            })
-                        })
-                        || file
-                            .get("violations")
-                            .and_then(|v| v.as_array())
-                            .is_some_and(|v| !v.is_empty())
-                })
-            });
-        if has_violations && hints.len() < MAX_HINTS {
-            // Preserve --rule / --rule-prefix / --fix-rule from the current
-            // invocation so the suggested preview doesn't widen scope.
-            hints.push(Hint::new(
-                "Preview auto-fixes",
-                build_lint_with_filter_flags(ctx, &["lint", "--fix", "--dry-run"]),
-            ));
-        }
-    } else if is_dry_run {
-        // Dry-run: if there are fixes that would be applied, suggest actually applying.
-        let total_fixed = data
-            .get("total_fixed")
-            .and_then(serde_json::Value::as_u64)
-            .unwrap_or(0);
-        // Also check old-shape `fixes` for backward compat.
-        let has_fixes = total_fixed > 0
-            || data
-                .get("fixes")
-                .and_then(|f| f.as_array())
-                .is_some_and(|a| !a.is_empty());
-        if has_fixes && hints.len() < MAX_HINTS {
-            // Apply hint mirrors the previewed scope — preserve --rule,
-            // --rule-prefix, and --fix-rule via the lint-aware builder.
-            hints.push(Hint::new(
-                "Apply auto-fixes",
-                build_lint_with_filter_flags(ctx, &["lint", "--fix"]),
-            ));
-        }
-    }
-    // If is_fix_mode && !is_dry_run: don't suggest fix hints (already applied).
-
-    // -----------------------------------------------------------------------
-    // UX-1: per-rule hints for HYALO001 and HYALO002.
-    // -----------------------------------------------------------------------
-    // Find the "worst-offender" file (first in files array — already sorted by total desc).
-    let worst_file = data
-        .get("files")
-        .and_then(|f| f.as_array())
-        .and_then(|arr| arr.first())
-        .and_then(|f| f.get("file"))
-        .and_then(serde_json::Value::as_str);
-
-    let rule_totals = collect_rule_totals(data);
-    let mut per_rule_hint_rules: Vec<String> = Vec::new();
-    for rule_id in &["HYALO001", "HYALO002"] {
-        if rule_totals.get(*rule_id).is_some_and(|&c| c > 0) {
-            per_rule_hint_rules.push(rule_id.to_string());
-        }
-    }
-    for rule_id in &per_rule_hint_rules {
-        if hints.len() >= MAX_HINTS {
-            break;
-        }
-        // De-dupe: don't add if we already have a hint with this rule.
-        let already = hints.iter().any(|h| h.cmd.contains(rule_id.as_str()));
-        if already {
-            continue;
-        }
-        if let Some(hint) = per_rule_hint(ctx, rule_id, worst_file) {
-            hints.push(hint);
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // UX-2: rule dominance hint.
-    // -----------------------------------------------------------------------
-    let grand_total: usize = rule_totals.values().sum();
-    if grand_total > 0
-        && hints.len() < MAX_HINTS
-        && let Some((dominant_rule, dominant_count_ref)) =
-            rule_totals.iter().max_by_key(|(_, c)| *c)
-    {
-        let dominant_count = *dominant_count_ref;
-        #[allow(clippy::cast_precision_loss)]
-        let ratio = dominant_count as f64 / grand_total as f64;
-        if ratio >= RULE_DOMINANCE_RATIO && dominant_count >= RULE_DOMINANCE_MIN {
-            let already = hints.iter().any(|h| {
-                h.cmd.contains("lint-rules show") && h.cmd.contains(dominant_rule.as_str())
-            });
-            if !already {
-                hints.push(Hint::new(
-                    format!("Tune {dominant_rule} if too noisy on this KB"),
-                    build_command_no_glob(ctx, &["lint-rules", "show", dominant_rule]),
-                ));
-            }
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Parse-error hint.
-    // -----------------------------------------------------------------------
-    let has_parse_errors = data
-        .get("files")
-        .and_then(|f| f.as_array())
-        .is_some_and(|files| {
-            files.iter().any(|file| {
-                file.get("rule_groups")
-                    .and_then(|rg| rg.as_array())
-                    .is_some_and(|groups| {
-                        groups.iter().any(|g| {
-                            g.get("violations")
-                                .and_then(|v| v.as_array())
-                                .is_some_and(|vs| {
-                                    vs.iter().any(|v| {
-                                        v.get("message")
-                                            .and_then(|m| m.as_str())
-                                            .is_some_and(|m| m.starts_with(PARSE_ERROR_PREFIX))
-                                    })
-                                })
-                        })
-                    })
-                    || file
-                        .get("violations")
-                        .and_then(|v| v.as_array())
-                        .is_some_and(|v| {
-                            v.iter().any(|violation| {
-                                violation
-                                    .get("message")
-                                    .and_then(|m| m.as_str())
-                                    .is_some_and(|m| m.starts_with(PARSE_ERROR_PREFIX))
-                            })
-                        })
-            })
-        });
-    if has_parse_errors && hints.len() < MAX_HINTS {
-        hints.push(Hint::new(
-            "Show all files with unfixable frontmatter errors",
-            build_command_with_glob_and_files(ctx, &["lint", "--limit", "0"]),
-        ));
-    }
-
-    // -----------------------------------------------------------------------
-    // SCHEMA → `types show <T>` (iter-143). Surface a per-type hint when
-    // SCHEMA violations land on files that declared a `type:`. Skip when
-    // the user is already focused on schema rules (--rule SCHEMA or
-    // --rule-prefix HYALO).
-    // -----------------------------------------------------------------------
-    let already_schema_focused = ctx.lint_rule.as_deref() == Some("SCHEMA")
-        || ctx
-            .lint_rule_prefix
-            .as_deref()
-            .is_some_and(|p| p.starts_with("HYALO"));
-    if !already_schema_focused && let Some(files) = data.get("files").and_then(|f| f.as_array()) {
-        // Collect distinct types that have at least one SCHEMA violation.
-        // Preserve first-seen order; cap at 2 distinct types to avoid noise.
-        //
-        // Inspect both the read-only mode shape (`rule_groups`) and the
-        // fix-mode shape (`remaining_groups`) so the hint fires regardless
-        // of which lint mode produced the output.
-        let mut schema_types: Vec<String> = Vec::new();
-        for file in files {
-            let has_schema_in = |key: &str| {
-                file.get(key)
-                    .and_then(|rg| rg.as_array())
-                    .is_some_and(|groups| {
-                        groups.iter().any(|g| {
-                            g.get("rule").and_then(serde_json::Value::as_str) == Some("SCHEMA")
-                        })
-                    })
-            };
-            if !has_schema_in("rule_groups") && !has_schema_in("remaining_groups") {
-                continue;
-            }
-            let Some(t) = file.get("type").and_then(serde_json::Value::as_str) else {
-                continue;
-            };
-            if !schema_types.iter().any(|x| x == t) {
-                schema_types.push(t.to_owned());
-            }
-            if schema_types.len() >= 2 {
-                break;
-            }
-        }
-        for t in &schema_types {
-            if hints.len() >= MAX_HINTS {
-                break;
-            }
-            hints.push(Hint::new(
-                format!("Show schema for type: {t}"),
-                build_command_no_glob(ctx, &["types", "show", t]),
-            ));
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Always suggest listing defined types.
-    // -----------------------------------------------------------------------
-    if hints.len() < MAX_HINTS {
-        hints.push(Hint::new(
-            "See defined type schemas",
-            build_command_no_glob(ctx, &["types", "list"]),
-        ));
-    }
-
-    hints
-}
-
-/// Drill-downs for `hyalo views list` (iter-210, dogfood UX-4).
-///
-/// The listing was a dead end: it named saved queries without saying how to
-/// run one, and on an empty vault it printed nothing at all with no way
-/// forward. The hints now always lead somewhere — into the first view when one
-/// exists, into creating one when none does.
-fn hints_for_views_list(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let names: Vec<&str> = data
-        .as_array()
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.get("name").and_then(serde_json::Value::as_str))
-                .collect()
-        })
-        .unwrap_or_default();
-
-    if let Some(first) = names.first() {
-        hints.push(Hint::new(
-            format!("Run the '{first}' view"),
-            build_command_no_glob(ctx, &["find", "--view", first]),
-        ));
-        if hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                format!("Delete the '{first}' view"),
-                build_command_no_glob(ctx, &["views", "remove", first]),
-            ));
-        }
-    } else {
-        // No views yet. `views set` needs at least one filter to be a valid
-        // command, so the suggestion carries a concrete (and runnable) one.
-        hints.push(Hint::new(
-            "Save a query as a view",
-            build_command_no_glob(
-                ctx,
-                &["views", "set", "drafts", "--property", "status=draft"],
-            ),
-        ));
-        hints.push(Hint::new(
-            "Survey the vault to decide which query is worth saving",
-            build_command_no_glob(ctx, &["summary"]),
-        ));
-    }
-
-    hints
-}
-
-/// Drill-downs for `hyalo lint-rules list` (iter-210, dogfood UX-4).
-///
-/// Picks the first *overridden* rule when there is one — that is the row a
-/// reader is most likely to be checking — and otherwise the first listed rule.
-fn hints_for_lint_rules_list(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let rules = data.as_array().map(Vec::as_slice).unwrap_or_default();
-    let rule_id = |v: &serde_json::Value| {
-        v.get("id")
-            .and_then(serde_json::Value::as_str)
-            .map(str::to_owned)
-    };
-    let overridden = rules
-        .iter()
-        .find(|r| {
-            r.get("has_override")
-                .and_then(serde_json::Value::as_bool)
-                .unwrap_or(false)
-        })
-        .and_then(rule_id);
-    let focus = overridden
-        .clone()
-        .or_else(|| rules.first().and_then(rule_id));
-
-    if let Some(ref id) = focus {
-        hints.push(Hint::new(
-            format!("Show what {id} checks and how to configure it"),
-            build_command_no_glob(ctx, &["lint-rules", "show", id]),
-        ));
-    }
-
-    if let Some(ref id) = overridden {
-        if hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                format!("Drop the {id} override and go back to the default"),
-                build_command_no_glob(ctx, &["lint-rules", "remove", id]),
-            ));
-        }
-    } else if let Some(ref id) = focus
-        && hints.len() < MAX_HINTS
-    {
-        hints.push(Hint::new(
-            format!("Turn {id} off for this vault"),
-            build_command_no_glob(ctx, &["lint-rules", "set", id, "--enabled", "false"]),
-        ));
-    }
-
-    if hints.len() < MAX_HINTS {
-        match focus {
-            // Narrowing lint to the rule in question is the fastest way to see
-            // whether it actually fires here.
-            Some(ref id) => hints.push(Hint::new(
-                format!("Run just {id} against the vault"),
-                build_command_no_glob(ctx, &["lint", "--rule", id]),
-            )),
-            None => hints.push(Hint::new(
-                "Run the markdown lint rules against the vault",
-                build_command_no_glob(ctx, &["lint"]),
-            )),
-        }
-    }
-
-    hints
-}
-
-/// Drill-downs for `hyalo lint-rules show <ID>` (NEW-18, dogfood pre3).
-///
-/// Was a hint dead end despite inspecting one specific, actionable rule — the
-/// natural next steps are running lint scoped to just that rule, and either
-/// toggling it or dropping an existing override.
-fn hints_for_lint_rules_show(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let mut hints = Vec::new();
-
-    let id = data
-        .get("id")
-        .and_then(serde_json::Value::as_str)
-        .or(ctx.lint_rule.as_deref());
-    let Some(id) = id else {
-        return hints;
-    };
-
-    hints.push(Hint::new(
-        format!("Run just {id} against the vault"),
-        build_command_no_glob(ctx, &["lint", "--rule", id]),
-    ));
-
-    let has_override = data.get("override").is_some_and(|o| !o.is_null());
-    if has_override {
-        if hints.len() < MAX_HINTS {
-            hints.push(Hint::new(
-                format!("Drop the {id} override and go back to the default"),
-                build_command_no_glob(ctx, &["lint-rules", "remove", id]),
-            ));
-        }
-    } else if hints.len() < MAX_HINTS {
-        let effective_enabled = data
-            .get("effective_enabled")
-            .and_then(serde_json::Value::as_bool)
-            .unwrap_or(true);
-        if effective_enabled {
-            hints.push(Hint::new(
-                format!("Turn {id} off for this vault"),
-                build_command_no_glob(ctx, &["lint-rules", "set", id, "--enabled", "false"]),
-            ));
-        } else {
-            hints.push(Hint::new(
-                format!("Turn {id} on for this vault"),
-                build_command_no_glob(ctx, &["lint-rules", "set", id, "--enabled", "true"]),
-            ));
-        }
-    }
-
-    hints
-}
-
-fn hints_for_types(ctx: &HintContext, data: &serde_json::Value) -> Vec<Hint> {
-    let subcommand = match &ctx.source {
-        HintSource::Types { subcommand } => subcommand.as_deref().unwrap_or("list"),
-        _ => "list",
-    };
-
-    let mut hints = Vec::new();
-
-    match subcommand {
-        "list" => {
-            // Suggest showing the first listed type.
-            if let Some(first_type) = data
-                .as_array()
-                .and_then(|arr| arr.first())
-                .and_then(|entry| entry.get("type"))
-                .and_then(serde_json::Value::as_str)
-            {
-                hints.push(Hint::new(
-                    format!("Show schema for type: {first_type}"),
-                    build_command_no_glob(ctx, &["types", "show", first_type]),
-                ));
-            }
-            if hints.len() < MAX_HINTS {
-                hints.push(Hint::new(
-                    "Validate all files against schema",
-                    build_command_no_glob(ctx, &["lint"]),
-                ));
-            }
-        }
-        "show" => {
-            let type_name = data.get("type").and_then(serde_json::Value::as_str);
-            // Suggest scaffolding a new file of this type when the type
-            // declares any `required` properties. Without required fields,
-            // `hyalo new` would only emit a `type:` stub — low value.
-            let has_required = data
-                .get("required")
-                .and_then(serde_json::Value::as_array)
-                .is_some_and(|arr| !arr.is_empty());
-            if let Some(name) = type_name
-                && has_required
-                && hints.len() < MAX_HINTS
-            {
-                let placeholder = format!("path/to/new-{name}.md");
-                hints.push(Hint::new(
-                    format!("Scaffold a new file of type: {name}"),
-                    build_command_no_glob(ctx, &["new", "--type", name, "--file", &placeholder]),
-                ));
-            }
-            if hints.len() < MAX_HINTS {
-                hints.push(Hint::new(
-                    "Validate files against schema",
-                    build_command_no_glob(ctx, &["lint"]),
-                ));
-            }
-            if hints.len() < MAX_HINTS {
-                hints.push(Hint::new(
-                    "List all type schemas",
-                    build_command_no_glob(ctx, &["types", "list"]),
-                ));
-            }
-            if let Some(name) = type_name
-                && hints.len() < MAX_HINTS
-            {
-                let filter = format!("type={name}");
-                hints.push(Hint::new(
-                    format!("Find files of type: {name}"),
-                    build_command_no_glob(ctx, &["find", "--property", &filter]),
-                ));
-            }
-        }
-        "set" => {
-            let type_name = data.get("type").and_then(serde_json::Value::as_str);
-            if let Some(name) = type_name
-                && hints.len() < MAX_HINTS
-            {
-                hints.push(Hint::new(
-                    format!("Review updated schema: {name}"),
-                    build_command_no_glob(ctx, &["types", "show", name]),
-                ));
-            }
-            if hints.len() < MAX_HINTS {
-                hints.push(Hint::new(
-                    "Validate files against schema",
-                    build_command_no_glob(ctx, &["lint"]),
-                ));
-            }
-        }
-        _ => {}
-    }
-
-    hints
-}
-
-fn hints_for_new(ctx: &HintContext, file: &str) -> Vec<Hint> {
-    vec![Hint::new(
-        "Validate the new file and see placeholder violations",
-        build_command_no_glob(ctx, &["lint", "--file", file]),
-    )]
-}
+//
+// `hints.rs` had grown to 5,059 lines -- a review hotspot flagged by the
+// 2026-08-27 deep review. The generators are now split by the command family
+// they serve; this file keeps only the vocabulary they share (the [`Hint`]
+// type, [`HintContext`], the thresholds) and the dispatch above. Every item is
+// re-exported at the visibility it already had, so no caller outside this
+// module can tell the difference.
+
+mod command;
+mod config;
+mod find;
+mod index;
+mod links;
+mod lint;
+mod mutation;
+mod summary;
+mod util;
+
+pub use command::{HintBuilder, shell_quote, shorten_index_path_for_hint};
+use command::{
+    build_command_no_glob, build_command_with_file, build_command_with_glob,
+    build_command_with_glob_and_files, build_find_command_composing,
+    build_find_command_preserving_filters, build_find_command_with_pattern,
+};
+use config::{
+    hints_for_lint_rules_list, hints_for_lint_rules_show, hints_for_new, hints_for_types,
+    hints_for_views_list,
+};
+use find::{format_confidence, hints_for_find};
+use index::{hints_for_create_index, hints_for_drop_index, hints_for_okf_index, hints_for_okf_log};
+use links::{hints_for_links_auto, hints_for_links_fix};
+use lint::hints_for_lint;
+use mutation::{
+    hints_for_backlinks, hints_for_mutation, hints_for_mv, hints_for_read, hints_for_task_mutation,
+    hints_for_task_read,
+};
+use summary::{hints_for_properties_summary, hints_for_summary, hints_for_tags_summary};
+use util::{first_modified_file, status_priority};
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    fn ctx(source: HintSource) -> HintContext {
-        HintContext::new(source)
-    }
-
-    // -----------------------------------------------------------------------
-    // HintBuilder (ARCH-4, iter-225)
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn hint_builder_basic_serialization() {
-        let b = HintBuilder::cmd("lint")
-            .flag_value("--rule", "HYALO006")
-            .flag("--detailed");
-        assert_eq!(b.build(), "hyalo lint --rule HYALO006 --detailed");
-    }
-
-    #[test]
-    fn hint_builder_quotes_shell_specials() {
-        let b = HintBuilder::cmd("find").flag_value("--property", "status=todo now");
-        assert_eq!(b.build(), "hyalo find --property 'status=todo now'");
-    }
-
-    #[test]
-    fn hint_builder_subcommand_groups() {
-        let b = HintBuilder::cmd("task toggle").arg("todo.md").flag("--all");
-        assert_eq!(b.build(), "hyalo task toggle todo.md --all");
-        assert_eq!(b.argv(), &["hyalo", "task", "toggle", "todo.md", "--all"]);
-    }
-
-    /// The typed half of ARCH-4: a command assembled through `HintBuilder`
-    /// must be accepted by the *real* clap parser. This is the in-process
-    /// version of what `tests/e2e/hint_execution.rs` proves by spawning the
-    /// binary — the `tags --limit 0` drift (a hint that satisfied a substring
-    /// assertion but failed to run) is now a unit-test failure the moment it
-    /// is written.
-    #[test]
-    fn hint_builder_commands_parse() {
-        let cases: Vec<(String, Vec<String>)> = vec![
-            ("hyalo summary".to_owned(), vec![]),
-            ("hyalo types list".to_owned(), vec![]),
-            (
-                "hyalo lint".to_owned(),
-                vec![
-                    "--rule".to_owned(),
-                    "HYALO006".to_owned(),
-                    "--detailed".to_owned(),
-                ],
-            ),
-            (
-                "hyalo find".to_owned(),
-                vec![
-                    "--broken-links".to_owned(),
-                    "--strict".to_owned(),
-                    "--limit".to_owned(),
-                    "0".to_owned(),
-                ],
-            ),
-            (
-                "hyalo task toggle".to_owned(),
-                vec!["todo.md".to_owned(), "--all".to_owned()],
-            ),
-            (
-                "hyalo tags summary".to_owned(),
-                vec!["--limit".to_owned(), "0".to_owned()],
-            ),
-            (
-                "hyalo set".to_owned(),
-                vec![
-                    "--property".to_owned(),
-                    "status=done".to_owned(),
-                    "--file".to_owned(),
-                    "notes/a.md".to_owned(),
-                ],
-            ),
-            (
-                "hyalo create-index".to_owned(),
-                vec!["--dir".to_owned(), "/my/vault".to_owned()],
-            ),
-        ];
-        for (subcommand, extra) in cases {
-            let subcommand = subcommand.strip_prefix("hyalo ").unwrap_or(&subcommand);
-            let mut b = HintBuilder::cmd(subcommand);
-            for arg in &extra {
-                b = b.raw(arg);
-            }
-            let argv: Vec<String> = b.argv().to_vec();
-            <crate::cli::args::Cli as clap::Parser>::try_parse_from(&argv)
-                .unwrap_or_else(|e| panic!("hint does not parse: {argv:?}: {e}"));
-        }
-    }
-
-    /// ARCH-4 drift guard: no NEW hand-written `"hyalo …"` command strings in
-    /// non-test source. Every hint command must go through [`HintBuilder`] so
-    /// the argv stays shell-quoted and parseable. A failure here means a new
-    /// hand-assembled hint was added — build it with `HintBuilder::cmd(...)`
-    /// instead (the remaining matches below are prose/starts_with checks, not
-    /// commands, and are allow-listed).
-    #[test]
-    fn no_raw_hyalo_command_literals() {
-        let src_dir = std::path::Path::new(env!("CARGO_MANIFEST_DIR")).join("src");
-        let allow: &[(&str, &str)] = &[
-            // starts_with() checks, not literals of hint commands
-            ("cli/help.rs", "if !trimmed.starts_with(\"hyalo \") {"),
-            // prose in an error message, not a command
-            (
-                "commands/mod.rs",
-                "\"hyalo does not support dotted path syntax for nested properties — --property \\",
-            ),
-            // warn.rs messages name the program, they are not commands
-            (
-                "warn.rs",
-                "\"hyalo is configured with dir = \\\"{dir_display}\\\".\\n  \\",
-            ),
-        ];
-        let mut offenders = Vec::new();
-        let mut stack = vec![src_dir.clone()];
-        while let Some(dir) = stack.pop() {
-            let entries = std::fs::read_dir(&dir)
-                .unwrap_or_else(|e| panic!("read_dir {}: {e}", dir.display()));
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.is_dir() {
-                    stack.push(path);
-                    continue;
-                }
-                if path.extension().and_then(std::ffi::OsStr::to_str) != Some("rs") {
-                    continue;
-                }
-                let rel = path
-                    .strip_prefix(&src_dir)
-                    .unwrap_or(&path)
-                    .to_string_lossy()
-                    .replace('\\', "/");
-                let text = std::fs::read_to_string(&path).unwrap_or_default();
-                // Skip test modules — their fixtures legitimately quote commands.
-                let text = match text.find("#[cfg(test)]") {
-                    Some(i) => &text[..i],
-                    None => &text[..],
-                };
-                for (n, line) in text.lines().enumerate() {
-                    let t = line.trim_start();
-                    if t.starts_with("//") || t.starts_with("//!") {
-                        continue;
-                    }
-                    if line.contains("\"hyalo ") {
-                        let key = (rel.as_str(), t);
-                        if !allow.contains(&key) {
-                            offenders.push(format!("{}:{}: {}", rel, n + 1, t));
-                        }
-                    }
-                }
-            }
-        }
-        assert!(
-            offenders.is_empty(),
-            "raw hyalo command string literals found (build hint commands with \
-             HintBuilder::cmd(...) instead — ARCH-4, iter-225):\n{}",
-            offenders.join("\n")
-        );
-    }
-
-    fn ctx_with_dir(source: HintSource, dir: &str) -> HintContext {
-        let mut ctx = HintContext::new(source);
-        ctx.dir = Some(dir.to_owned());
-        ctx
-    }
-
-    fn ctx_with_glob(source: HintSource, glob: &str) -> HintContext {
-        let mut ctx = HintContext::new(source);
-        ctx.glob = vec![glob.to_owned()];
-        ctx
-    }
-
-    // --- global flag placement (iter-213, dogfood UX-5) ---
-
-    /// Every builder must leave `--dir`/`--format` at the tail, so a block of
-    /// hints under one result set reads as variations of one command instead
-    /// of scattering the same two flags at different offsets.
-    #[test]
-    fn global_flags_are_the_last_thing_every_builder_pushes() {
-        let mut ctx = HintContext::new(HintSource::Find);
-        ctx.dir = Some("kb".to_owned());
-        ctx.format = Some("text".to_owned());
-        ctx.glob = vec!["notes/*.md".to_owned()];
-        ctx.file_targets = vec!["a.md".to_owned()];
-
-        for cmd in [
-            build_command_no_glob(&ctx, &["summary"]),
-            build_command_with_file(&ctx, &["read"], "a.md", &[]),
-            build_command_with_glob(&ctx, &["lint"]),
-            build_command_with_glob_and_files(&ctx, &["lint"]),
-            build_find_command_preserving_filters(&ctx, &["--limit", "5"]),
-            build_views_set_command(&ctx, "my-view"),
-        ] {
-            assert!(
-                cmd.ends_with("--dir kb --format text"),
-                "global flags must trail: {cmd}"
-            );
-        }
-    }
-
-    // --- shell_quote ---
-
-    #[test]
-    fn shell_quote_plain_string() {
-        assert_eq!(shell_quote("status"), "status");
-    }
-
-    #[test]
-    fn shell_quote_string_with_space() {
-        assert_eq!(shell_quote("in progress"), "'in progress'");
-    }
-
-    #[test]
-    fn shell_quote_string_with_special_chars() {
-        assert_eq!(shell_quote("foo$bar"), "'foo$bar'");
-    }
-
-    #[test]
-    fn shell_quote_string_with_single_quote() {
-        assert_eq!(shell_quote("it's"), "'it'\\''s'");
-    }
-
-    #[test]
-    fn shell_quote_glob_chars() {
-        assert_eq!(shell_quote("**/*.md"), "'**/*.md'");
-    }
-
-    #[test]
-    fn shell_quote_empty_string() {
-        assert_eq!(shell_quote(""), "''");
-    }
-
-    // --- build_command ---
-
-    #[test]
-    fn build_command_no_flags() {
-        let c = ctx(HintSource::Summary);
-        assert_eq!(
-            build_command_no_glob(&c, &["properties"]),
-            "hyalo properties"
-        );
-    }
-
-    #[test]
-    fn build_command_with_dir() {
-        let c = ctx_with_dir(HintSource::Summary, "/my/vault");
-        assert_eq!(
-            build_command_no_glob(&c, &["tags"]),
-            "hyalo tags --dir /my/vault"
-        );
-    }
-
-    #[test]
-    fn build_command_with_glob_propagated() {
-        let c = ctx_with_glob(HintSource::PropertiesSummary, "**/*.md");
-        assert_eq!(
-            build_command_with_glob(&c, &["properties"]),
-            "hyalo properties --glob '**/*.md'"
-        );
-    }
-
-    // --- status_priority ---
-
-    #[test]
-    fn status_priority_ordering() {
-        assert!(status_priority("in-progress") < status_priority("planned"));
-        assert!(status_priority("planned") < status_priority("draft"));
-        assert!(status_priority("draft") < status_priority("custom"));
-        assert!(status_priority("custom") < status_priority("completed"));
-    }
-
-    // --- hints_for_summary ---
-
-    #[test]
-    fn summary_always_includes_properties_and_tags() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 10, "by_directory": []},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [],
-            "tasks": {"total": 0, "done": 0},
-            "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(hints.iter().any(|h| {
-            h.cmd == "hyalo properties"
-                || (h.cmd.starts_with("hyalo properties ") && h.cmd.contains("--dir "))
-                || (h.cmd.starts_with("hyalo properties ") && h.cmd.contains("--glob "))
-        }));
-        assert!(hints.iter().any(|h| {
-            h.cmd == "hyalo tags"
-                || (h.cmd.starts_with("hyalo tags ") && h.cmd.contains("--dir "))
-                || (h.cmd.starts_with("hyalo tags ") && h.cmd.contains("--glob "))
-        }));
-    }
-
-    #[test]
-    fn summary_suggests_tasks_todo_when_open_tasks() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 5, "by_directory": []},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [],
-            "tasks": {"total": 10, "done": 3},
-            "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("find")
-                && h.cmd.contains("--task")
-                && h.cmd.contains("todo"))
-        );
-    }
-
-    #[test]
-    fn summary_omits_tasks_todo_when_all_done() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 5, "by_directory": []},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [],
-            "tasks": {"total": 10, "done": 10},
-            "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(!hints.iter().any(|h| h.cmd.contains("--todo")));
-    }
-
-    #[test]
-    fn summary_picks_interesting_status_values() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 5, "by_directory": []},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [
-                {"value": "completed", "files": ["a.md"]},
-                {"value": "in-progress", "files": ["b.md"]},
-                {"value": "planned", "files": ["c.md"]}
-            ],
-            "tasks": {"total": 0, "done": 0},
-            "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        // in-progress should appear before completed
-        let in_progress_pos = hints.iter().position(|h| h.cmd.contains("in-progress"));
-        let completed_pos = hints.iter().position(|h| h.cmd.contains("completed"));
-        assert!(in_progress_pos.is_some(), "should suggest in-progress");
-        // completed may appear (only if limit not reached) or not — but in-progress must come first
-        if let Some(cp) = completed_pos {
-            assert!(in_progress_pos.unwrap() < cp);
-        }
-    }
-
-    #[test]
-    fn summary_max_hints_not_exceeded() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 5, "by_directory": []},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [
-                {"value": "in-progress", "files": ["a.md"]},
-                {"value": "planned", "files": ["b.md"]},
-                {"value": "draft", "files": ["c.md"]},
-                {"value": "idea", "files": ["d.md"]}
-            ],
-            "tasks": {"total": 5, "done": 1},
-            "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(hints.len() <= MAX_HINTS);
-    }
-
-    // --- hints_for_properties_summary ---
-
-    #[test]
-    fn properties_summary_top3_by_count() {
-        let c = ctx(HintSource::PropertiesSummary);
-        let data = json!([
-            {"name": "title", "type": "text", "count": 100},
-            {"name": "status", "type": "text", "count": 50},
-            {"name": "tags", "type": "list", "count": 30},
-            {"name": "author", "type": "text", "count": 5}
-        ]);
-        let hints = generate_hints(&c, &data, None);
-        assert_eq!(hints.len(), 3);
-        assert!(hints[0].cmd.contains("title"));
-        assert!(hints[1].cmd.contains("status"));
-        assert!(hints[2].cmd.contains("tags"));
-        // author should not appear (rank 4)
-        assert!(!hints.iter().any(|h| h.cmd.contains("author")));
-    }
-
-    #[test]
-    fn properties_summary_empty_data() {
-        let c = ctx(HintSource::PropertiesSummary);
-        let hints = generate_hints(&c, &json!([]), None);
-        assert!(hints.is_empty());
-    }
-
-    #[test]
-    fn properties_summary_propagates_glob() {
-        let c = ctx_with_glob(HintSource::PropertiesSummary, "notes/*.md");
-        let data = json!([{"name": "status", "type": "text", "count": 5}]);
-        let hints = generate_hints(&c, &data, None);
-        assert!(hints[0].cmd.contains("--glob"));
-        assert!(hints[0].cmd.contains("notes/*.md"));
-    }
-
-    // --- hints_for_find ---
-
-    fn make_find_item(file: &str, status: Option<&str>, tags: &[&str]) -> serde_json::Value {
-        let mut props = serde_json::Map::new();
-        if let Some(s) = status {
-            props.insert("status".to_owned(), serde_json::Value::String(s.to_owned()));
-        }
-        json!({
-            "file": file,
-            "properties": props,
-            "tags": tags,
-            "sections": [],
-            "tasks": [],
-            "links": [],
-            "modified": "2026-01-01T00:00:00Z"
-        })
-    }
-
-    #[test]
-    fn find_empty_results_no_hints() {
-        let c = ctx(HintSource::Find);
-        let hints = generate_hints(&c, &json!([]), None);
-        assert!(hints.is_empty());
-    }
-
-    #[test]
-    fn find_single_result_suggests_read_and_backlinks() {
-        let c = ctx(HintSource::Find);
-        let items = vec![make_find_item("notes/alpha.md", None, &[])];
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("read") && h.cmd.contains("alpha.md")),
-            "should suggest read: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("backlinks") && h.cmd.contains("alpha.md")),
-            "should suggest backlinks: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_many_results_suggests_top_tag() {
-        let c = ctx(HintSource::Find);
-        // 6 results; rust appears 4 times, cli 2 times — rust should be suggested.
-        let items = vec![
-            make_find_item("a.md", Some("planned"), &["rust", "cli"]),
-            make_find_item("b.md", Some("planned"), &["rust"]),
-            make_find_item("c.md", Some("in-progress"), &["rust"]),
-            make_find_item("d.md", Some("completed"), &["rust"]),
-            make_find_item("e.md", Some("completed"), &["cli"]),
-            make_find_item("f.md", Some("completed"), &[]),
-        ];
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("--tag") && h.cmd.contains("rust")),
-            "should suggest --tag rust (most common): {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_many_results_suggests_interesting_status() {
-        let c = ctx(HintSource::Find);
-        // 6 results; in-progress is more interesting than completed.
-        let items = vec![
-            make_find_item("a.md", Some("in-progress"), &[]),
-            make_find_item("b.md", Some("completed"), &[]),
-            make_find_item("c.md", Some("completed"), &[]),
-            make_find_item("d.md", Some("completed"), &[]),
-            make_find_item("e.md", Some("completed"), &[]),
-            make_find_item("f.md", Some("completed"), &[]),
-        ];
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("--property") && h.cmd.contains("status=in-progress")),
-            "should prefer in-progress status: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_many_results_no_tags_falls_back_to_status() {
-        let c = ctx(HintSource::Find);
-        // 6 results, none with tags; should still suggest status narrowing.
-        let items = vec![
-            make_find_item("a.md", Some("planned"), &[]),
-            make_find_item("b.md", Some("planned"), &[]),
-            make_find_item("c.md", Some("planned"), &[]),
-            make_find_item("d.md", Some("planned"), &[]),
-            make_find_item("e.md", Some("planned"), &[]),
-            make_find_item("f.md", Some("planned"), &[]),
-        ];
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("--property") && h.cmd.contains("status=planned")),
-            "should suggest status filter: {hints:?}"
-        );
-        // No --tag hints when no tags exist.
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("--tag")),
-            "should not suggest --tag when no tags: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_hints_never_exceed_max() {
-        let c = ctx(HintSource::Find);
-        // 10 results with varied tags and statuses.
-        let items: Vec<serde_json::Value> = (0..10)
-            .map(|i| make_find_item(&format!("{i}.md"), Some("planned"), &["rust", "cli"]))
-            .collect();
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(hints.len() <= MAX_HINTS);
-    }
-
-    #[test]
-    fn find_sort_hint_preserves_existing_filters() {
-        let mut c = ctx(HintSource::Find);
-        c.property_filters = vec!["status=draft".to_owned()];
-        c.tag_filters = vec!["research".to_owned()];
-        // 6 results to trigger sort/limit hints.
-        let items: Vec<serde_json::Value> = (0..6)
-            .map(|i| make_find_item(&format!("{i}.md"), Some("draft"), &["research"]))
-            .collect();
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        let sort_hint = hints.iter().find(|h| h.cmd.contains("--sort"));
-        assert!(sort_hint.is_some(), "should include a sort hint: {hints:?}");
-        let cmd = &sort_hint.unwrap().cmd;
-        assert!(
-            cmd.contains("--property status=draft"),
-            "sort hint should preserve --property filter: {cmd}"
-        );
-        assert!(
-            cmd.contains("--tag research"),
-            "sort hint should preserve --tag filter: {cmd}"
-        );
-    }
-
-    #[test]
-    fn find_limit_hint_preserves_existing_filters() {
-        let mut c = ctx(HintSource::Find);
-        c.tag_filters = vec!["iteration".to_owned()];
-        let items: Vec<serde_json::Value> = (0..6)
-            .map(|i| make_find_item(&format!("{i}.md"), Some("planned"), &["iteration"]))
-            .collect();
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        let limit_hint = hints.iter().find(|h| h.cmd.contains("--limit"));
-        assert!(
-            limit_hint.is_some(),
-            "should include a limit hint: {hints:?}"
-        );
-        let cmd = &limit_hint.unwrap().cmd;
-        assert!(
-            cmd.contains("--tag iteration"),
-            "limit hint should preserve --tag filter: {cmd}"
-        );
-    }
-
-    // --- flag propagation ---
-
-    #[test]
-    fn dir_flag_propagated_to_all_hints() {
-        let c = ctx_with_dir(HintSource::TagsSummary, "/vault");
-        // tags summary returns a bare array [{name, count}, ...]
-        let data = json!([{"name": "rust", "count": 5}]);
-        let hints = generate_hints(&c, &data, None);
-        assert!(hints[0].cmd.contains("--dir"));
-        assert!(hints[0].cmd.contains("/vault"));
-    }
-
-    // --- new generator tests ---
-
-    #[test]
-    fn mutation_hints_suggest_verify_and_read() {
-        let c = ctx(HintSource::Set);
-        let data = json!({
-            "property": "status",
-            "value": "completed",
-            "modified": ["notes/alpha.md"],
-            "skipped": [],
-            "total": 1
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("find") && h.cmd.contains("alpha.md")),
-            "should suggest verify: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("read") && h.cmd.contains("alpha.md")),
-            "should suggest read: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn read_hints_suggest_metadata_and_backlinks() {
-        let c = ctx(HintSource::Read);
-        let data = json!({"file": "notes/alpha.md", "content": "Some content"});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("find") && h.cmd.contains("alpha.md")),
-            "should suggest find: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("backlinks") && h.cmd.contains("alpha.md")),
-            "should suggest backlinks: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn backlinks_hints_suggest_read_and_outgoing() {
-        let c = ctx(HintSource::Backlinks);
-        let data = json!({
-            "file": "target.md",
-            "backlinks": [{"source": "a.md", "line": 5, "target": "target"}],
-            "total": 1
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("read") && h.cmd.contains("target.md")),
-            "should suggest read target: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("read") && h.cmd.contains("a.md")),
-            "should suggest read first backlink source: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn create_index_hints_suggest_find_and_drop() {
-        let c = ctx(HintSource::CreateIndex);
-        let data = json!({"path": ".hyalo-index", "files_indexed": 42, "warnings": 0});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("find") && h.cmd.contains("--index")),
-            "should suggest find with index: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("drop-index")),
-            "should suggest drop-index: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn drop_index_hints_suggest_create() {
-        let c = ctx(HintSource::DropIndex);
-        let data = json!({"deleted": ".hyalo-index"});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("create-index")),
-            "should suggest create-index: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn mv_dry_run_hints_suggest_apply() {
-        let c = ctx(HintSource::Mv);
-        let data = json!({
-            "from": "old.md",
-            "to": "new.md",
-            "dry_run": true,
-            "updated_files": [],
-            "total_files_updated": 0,
-            "total_links_updated": 0
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("mv")
-                && h.cmd.contains("new.md")
-                && !h.cmd.contains("dry-run")),
-            "should suggest applying the move: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn mv_applied_hints_suggest_read_and_backlinks() {
-        let c = ctx(HintSource::Mv);
-        let data = json!({
-            "from": "old.md",
-            "to": "new.md",
-            "dry_run": false,
-            "updated_files": [],
-            "total_files_updated": 0,
-            "total_links_updated": 0
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("read") && h.cmd.contains("new.md")),
-            "should suggest reading moved file: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("backlinks") && h.cmd.contains("new.md")),
-            "should suggest checking backlinks: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn task_read_undone_suggests_toggle() {
-        let c = ctx(HintSource::TaskRead);
-        let data =
-            json!({"file": "todo.md", "line": 5, "status": " ", "text": "Fix bug", "done": false});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("task toggle")),
-            "should suggest toggling undone task: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn task_read_done_omits_toggle() {
-        let c = ctx(HintSource::TaskRead);
-        let data =
-            json!({"file": "todo.md", "line": 5, "status": "x", "text": "Fix bug", "done": true});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("task toggle")),
-            "should not suggest toggling already-done task: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("--task todo")),
-            "should suggest viewing open tasks: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn task_mutation_hints_suggest_remaining_tasks() {
-        let c = ctx(HintSource::TaskToggle);
-        let data =
-            json!({"file": "todo.md", "line": 5, "status": "x", "text": "Fix bug", "done": true});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("find")
-                && h.cmd.contains("--task")
-                && h.cmd.contains("todo")),
-            "should suggest finding remaining tasks: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn links_fix_dry_run_hints_suggest_apply() {
-        let c = ctx(HintSource::LinksFix);
-        let data = json!({
-            "broken": 5,
-            "fixable": 3,
-            "unfixable": 2,
-            "applied": false,
-            "fixes": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("links fix --apply")),
-            "should suggest applying fixes: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("--broken-links")),
-            "should suggest finding broken links: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn links_fix_apply_failures_suggest_checking_failed_fixes() {
-        let c = ctx(HintSource::LinksFix);
-        let data = json!({
-            "broken": 3,
-            "fixable": 3,
-            "unfixable": 0,
-            "applied": true,
-            "fixes": [],
-            "failed": 2,
-            "failed_fixes": [],
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.is_empty() && h.description.contains("failed_fixes")),
-            "should surface an advice-only hint pointing at failed_fixes: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn links_fix_no_failures_omits_failed_fixes_hint() {
-        let c = ctx(HintSource::LinksFix);
-        let data = json!({
-            "broken": 3,
-            "fixable": 3,
-            "unfixable": 0,
-            "applied": true,
-            "fixes": [],
-            "failed": 0,
-            "failed_fixes": [],
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.description.contains("failed_fixes")),
-            "should not mention failed_fixes when nothing failed: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn links_auto_apply_failures_suggest_checking_apply_outcomes() {
-        let c = ctx(HintSource::LinksAuto);
-        let data = json!({
-            "matched": 5,
-            "dry_run": false,
-            "applied": true,
-            "files_applied": 3,
-            "files_skipped": 0,
-            "files_failed": 2,
-            "apply_outcomes": [],
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.is_empty() && h.description.contains("apply_outcomes")),
-            "should surface an advice-only hint pointing at apply_outcomes: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn links_auto_no_failures_omits_apply_outcomes_hint() {
-        let c = ctx(HintSource::LinksAuto);
-        let data = json!({
-            "matched": 5,
-            "dry_run": false,
-            "applied": true,
-            "files_applied": 5,
-            "files_skipped": 0,
-            "files_failed": 0,
-            "apply_outcomes": [],
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.description.contains("apply_outcomes")),
-            "should not mention apply_outcomes when nothing failed: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_broad_query_suggests_summary() {
-        let c = ctx(HintSource::Find);
-        // 15 results, no filters
-        let items: Vec<serde_json::Value> = (0..15)
-            .map(|i| make_find_item(&format!("{i}.md"), Some("completed"), &[]))
-            .collect();
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("summary")),
-            "broad query should suggest summary: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_with_filters_does_not_suggest_summary() {
-        let mut c = ctx(HintSource::Find);
-        c.tag_filters = vec!["rust".to_owned()];
-        let items: Vec<serde_json::Value> = (0..15)
-            .map(|i| make_find_item(&format!("{i}.md"), Some("completed"), &["rust"]))
-            .collect();
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("summary")),
-            "filtered query should not suggest summary: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_suppresses_already_filtered_tag() {
-        let mut c = ctx(HintSource::Find);
-        c.tag_filters = vec!["rust".to_owned()];
-        let items: Vec<serde_json::Value> = (0..10)
-            .map(|i| make_find_item(&format!("{i}.md"), Some("planned"), &["rust", "cli"]))
-            .collect();
-        let data = json!(items);
-        let hints = generate_hints(&c, &data, None);
-        // Should NOT suggest narrowing *by* the already-filtered `rust` tag.
-        // Narrow hints now compose with the active filter (BUG-8), so the
-        // preserved `--tag rust` legitimately appears in the command; the
-        // *narrow target* is what must differ. Assert the narrow hint's
-        // description names `cli`, not `rust`.
-        assert!(
-            !hints.iter().any(|h| h.description == "Narrow by tag: rust"
-                || h.description.starts_with("Narrow by tag: rust ")),
-            "should not suggest narrowing by already-filtered tag: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.description.starts_with("Narrow by tag: cli")),
-            "should suggest narrowing by non-filtered tag: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn summary_broken_links_suggests_links_fix() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": 10,
-            "links": {"total": 20, "broken": 3},
-            "properties": [],
-            "tags": [],
-            "status": [],
-            "tasks": {"total": 0, "done": 0},
-            "orphans": 0
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("links fix")),
-            "summary with broken links should suggest links fix: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("--broken-links")),
-            "summary with broken links should also suggest find --broken-links: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn summary_no_broken_links_omits_links_fix() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": 10,
-            "links": {"total": 20, "broken": 0},
-            "properties": [],
-            "tags": [],
-            "status": [],
-            "tasks": {"total": 0, "done": 0},
-            "orphans": 0
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("links fix")),
-            "summary without broken links should not suggest links fix: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_with_broken_links_suggests_links_fix() {
-        let c = ctx(HintSource::Find);
-        let item = json!({
-            "file": "doc.md",
-            "properties": {},
-            "tags": [],
-            "sections": [],
-            "tasks": [],
-            "links": [
-                {"target": "existing.md", "path": "existing.md", "kind": "wiki"},
-                {"target": "gone.md", "path": null, "kind": "wiki"}
-            ],
-            "modified": "2026-01-01T00:00:00Z"
-        });
-        let data = json!([item]);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("links fix")),
-            "find results with broken links should suggest links fix: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_without_broken_links_omits_links_fix() {
-        let c = ctx(HintSource::Find);
-        let item = json!({
-            "file": "doc.md",
-            "properties": {},
-            "tags": [],
-            "sections": [],
-            "tasks": [],
-            "links": [
-                {"target": "existing.md", "path": "existing.md", "kind": "wiki"}
-            ],
-            "modified": "2026-01-01T00:00:00Z"
-        });
-        let data = json!([item]);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("links fix")),
-            "find results without broken links should not suggest links fix: {hints:?}"
-        );
-    }
-
-    // --- hints_for_lint ---
-
-    #[test]
-    fn lint_hints_name_hidden_errors_when_truncated() {
-        // UX-4 (dogfood v0.20.0): the truncated `files[]` slice shows only
-        // warnings, but the authoritative count is `errors: 4` — the
-        // show-all hint must say the errors are hidden, or the summary line
-        // reads like a bug.
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{
-                "file": "warn-only.md",
-                "rule_groups": [{
-                    "rule": "MD040",
-                    "severity": "warn",
-                    "count": 1,
-                    "shown": 1,
-                    "violations": [{"severity": "warn", "message": "fenced code block should have a language"}]
-                }]
-            }],
-            "files_truncated": true,
-            "files_with_violations": 60,
-            "errors": 4,
-            "warnings": 7716
-        });
-        let hints = generate_hints(&c, &data, None);
-        let show_all = hints
-            .iter()
-            .find(|h| h.cmd.contains("--limit 0"))
-            .expect("show-all hint should be present");
-        assert!(
-            show_all.description.contains("4 errors hidden"),
-            "hint should name the hidden errors: {show_all:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_no_hidden_errors_claim_in_fix_mode_dry_run() {
-        // Review M-1 (PR #277): fix-mode shapes groups as
-        // `fixed_groups`/`remaining_groups`, so the read-only
-        // `rule_groups` computation must not claim hidden errors there —
-        // even though the truncation hint itself is kept in dry-run mode.
-        let mut c = ctx(HintSource::Lint);
-        c.dry_run = true;
-        c.lint_is_fix = true;
-        let data = json!({
-            "files": [{
-                "file": "err.md",
-                "remaining_groups": [{
-                    "rule": "MD011",
-                    "severity": "error",
-                    "count": 1,
-                    "violations": [{"severity": "error", "message": "Reversed link syntax: (a)[b]"}]
-                }]
-            }],
-            "files_truncated": true,
-            "files_with_violations": 60,
-            "errors": 4,
-            "warnings": 7716,
-            "dry_run": true,
-            "total_fixed": 0,
-            "total_remaining": 4
-        });
-        let hints = generate_hints(&c, &data, None);
-        let show_all = hints
-            .iter()
-            .find(|h| h.cmd.contains("--limit 0"))
-            .expect("show-all hint should still be present in dry-run mode");
-        assert!(
-            !show_all.description.contains("hidden"),
-            "fix-mode dry-run must not claim hidden errors: {show_all:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_suggest_fix_when_violations() {
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{"file": "test.md", "violations": [{"severity": "error", "message": "missing required property"}]}],
-            "violations": 1,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(!hints.is_empty());
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("lint --fix")),
-            "should suggest lint --fix: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_suggest_apply_when_dry_run() {
-        let mut c = ctx(HintSource::Lint);
-        c.dry_run = true;
-        c.lint_is_fix = true; // --dry-run requires --fix per CLI spec
-        let data = json!({
-            "files": [],
-            "violations": 0,
-            "total_fixed": 3,
-            "total_remaining": 0,
-            "fixes": [{"file": "test.md", "actions": [{"kind": "insert-default", "property": "status", "new": "draft"}]}],
-            "dry_run": true,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("lint --fix") && !h.cmd.contains("--dry-run")),
-            "dry-run mode should suggest lint --fix without --dry-run: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_always_suggest_types_list() {
-        let c = ctx(HintSource::Lint);
-        let data = json!({"files": [], "violations": 0});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("types list")),
-            "should always suggest types list: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_never_exceed_max() {
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{"file": "test.md", "violations": [{"severity": "error", "message": "x", "type": "iteration"}]}],
-            "violations": 5,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(hints.len() <= MAX_HINTS);
-    }
-
-    // --- hints_for_types ---
-
-    #[test]
-    fn types_list_hints_suggest_show() {
-        let c = ctx(HintSource::Types {
-            subcommand: Some("list".to_owned()),
-        });
-        let data = json!([
-            {"type": "iteration", "required": ["title"], "has_filename_template": true, "property_count": 3},
-            {"type": "note", "required": [], "has_filename_template": false, "property_count": 1},
-        ]);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("types show")),
-            "should suggest types show: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("lint")),
-            "should suggest lint: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn types_show_hints_suggest_lint_and_find() {
-        let c = ctx(HintSource::Types {
-            subcommand: Some("show".to_owned()),
-        });
-        let data = json!({"type": "iteration", "required": ["title"], "properties": {}});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("lint")),
-            "should suggest lint: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("find --property")),
-            "should suggest find --property: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn types_show_hints_suggest_scaffold_when_required_nonempty() {
-        // iter-143: when the type declares any `required` properties, `types
-        // show` surfaces a hint to scaffold a new file via `hyalo new`.
-        let c = ctx(HintSource::Types {
-            subcommand: Some("show".to_owned()),
-        });
-        let data = json!({
-            "type": "iteration",
-            "required": ["title", "status"],
-            "properties": {},
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("new --type iteration")),
-            "should suggest scaffolding a new file: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn types_show_hints_no_scaffold_when_required_empty() {
-        // iter-143: when `required` is empty, the scaffold hint is dropped
-        // (it would only emit a `type:` stub — low value).
-        let c = ctx(HintSource::Types {
-            subcommand: Some("show".to_owned()),
-        });
-        let data = json!({
-            "type": "note",
-            "required": [],
-            "properties": {},
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("new --type")),
-            "should NOT suggest scaffolding when required is empty: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_schema_violation_suggests_types_show() {
-        // iter-143: when SCHEMA violations land on a typed file, surface
-        // `hyalo types show <T>`.
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{
-                "file": "foo.md",
-                "type": "iteration",
-                "rule_groups": [{
-                    "rule": "SCHEMA", "count": 2, "shown": 2,
-                    "truncated": false, "severity": "error", "autofixable": false,
-                    "violations": [],
-                }]
-            }],
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("types show iteration")),
-            "should suggest types show for the failing type: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_schema_violation_suggests_types_show_in_fix_mode() {
-        // iter-143 follow-up (Copilot review on PR #169): the SCHEMA→`types
-        // show` hint must also fire in `--fix` / `--fix --dry-run` output,
-        // where violations live under `remaining_groups` instead of
-        // `rule_groups`.
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{
-                "file": "foo.md",
-                "type": "iteration",
-                "fixed_groups": [],
-                "remaining_groups": [{
-                    "rule": "SCHEMA", "count": 1, "shown": 1,
-                    "truncated": false, "severity": "error", "autofixable": false,
-                    "violations": [],
-                }],
-                "conflicts": [],
-            }],
-            "dry_run": true,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("types show iteration")),
-            "should suggest types show in fix-mode too: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_schema_violation_skipped_when_already_focused() {
-        // iter-143: when the user is already filtering on SCHEMA via --rule
-        // SCHEMA (or --rule-prefix HYALO), the `types show` hint would be
-        // redundant — suppress it.
-        let mut c = ctx(HintSource::Lint);
-        c.lint_rule = Some("SCHEMA".to_owned());
-        let data = json!({
-            "files": [{
-                "file": "foo.md",
-                "type": "iteration",
-                "rule_groups": [{
-                    "rule": "SCHEMA", "count": 1, "shown": 1,
-                    "truncated": false, "severity": "error", "autofixable": false,
-                    "violations": [],
-                }]
-            }],
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("types show")),
-            "should NOT suggest types show when --rule SCHEMA: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn files_from_hints_fire_on_missing_and_outside_vault() {
-        // iter-143: the FilesFromCounterSummary path produces advice hints.
-        let c = ctx(HintSource::Find);
-        let data = json!({"results": [], "total": 0});
-        let counters = FilesFromCounterSummary {
-            files_missing: 3,
-            files_skipped_outside_vault: 1,
-        };
-        let hints = generate_hints_with_counters(&c, &data, None, Some(counters));
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.is_empty() && h.description.contains("3 input path")),
-            "should warn about missing inputs: {hints:?}"
-        );
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.is_empty() && h.description.contains("outside the vault")),
-            "should warn about outside-vault inputs: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn files_from_hints_silent_when_zero_counters() {
-        let c = ctx(HintSource::Find);
-        let data = json!({"results": []});
-        let counters = FilesFromCounterSummary::default();
-        let hints = generate_hints_with_counters(&c, &data, None, Some(counters));
-        assert!(
-            !hints.iter().any(|h| h.cmd.is_empty()),
-            "no advice hints expected when counters are zero: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn types_set_hints_suggest_show_and_lint() {
-        let c = ctx(HintSource::Types {
-            subcommand: Some("set".to_owned()),
-        });
-        let data = json!({"type": "iteration", "action": "updated"});
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("types show iteration")),
-            "should suggest types show for updated type: {hints:?}"
-        );
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("lint")),
-            "should suggest lint: {hints:?}"
-        );
-    }
-
-    // --- UX-1: per-rule hints for HYALO001 / HYALO002 ---
-
-    #[test]
-    fn lint_hints_hyalo001_suggests_fix_rule() {
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{
-                "file": "test.md",
-                "rule_groups": [{"rule": "HYALO001", "count": 3, "shown": 3, "truncated": false,
-                                 "severity": "error", "autofixable": true,
-                                 "violations": [{"line": 4, "column": 1, "message": "bare []"}]}]
-            }],
-            "total": 3,
-            "rules_fired": 1,
-            "files_with_violations": 1,
-            "files_checked": 1,
-            "files_truncated": false,
-            "errors": 3,
-            "warnings": 0,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("HYALO001") && h.cmd.contains("--fix")),
-            "should suggest lint --rule HYALO001 --fix for HYALO001 violations: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_hyalo002_suggests_find_todo() {
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{
-                "file": "iterations/iter-1.md",
-                "rule_groups": [{"rule": "HYALO002", "count": 5, "shown": 3, "truncated": true,
-                                 "severity": "error", "autofixable": false,
-                                 "violations": [{"line": 21, "column": 1, "message": "completed but tasks remain"}]}]
-            }],
-            "total": 5,
-            "rules_fired": 1,
-            "files_with_violations": 1,
-            "files_checked": 1,
-            "files_truncated": false,
-            "errors": 5,
-            "warnings": 0,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("find --task todo") && h.cmd.contains("iter-1")),
-            "should suggest find --task todo with worst-offender file: {hints:?}"
-        );
-    }
-
-    // --- UX-2: rule dominance hint ---
-
-    #[test]
-    fn lint_hints_dominant_rule_suggests_tune() {
-        let c = ctx(HintSource::Lint);
-        // MD013 has 80 of 100 total violations → 80% share, ≥50 absolute.
-        let mut groups: Vec<serde_json::Value> = Vec::new();
-        for _ in 0..80 {
-            groups.push(json!({"line": 1, "column": 1, "message": "line too long"}));
-        }
-        let data = json!({
-            "files": [
-                {"file": "a.md", "rule_groups": [
-                    {"rule": "MD013", "count": 80, "shown": 3, "truncated": true,
-                     "severity": "warn", "autofixable": false,
-                     "violations": [{"line": 1, "column": 1, "message": "line too long"}]}
-                ]},
-                {"file": "b.md", "rule_groups": [
-                    {"rule": "HYALO001", "count": 20, "shown": 3, "truncated": true,
-                     "severity": "error", "autofixable": true,
-                     "violations": [{"line": 2, "column": 1, "message": "bare []"}]}
-                ]}
-            ],
-            "total": 100,
-            "rules_fired": 2,
-            "files_with_violations": 2,
-            "files_checked": 5,
-            "files_truncated": false,
-            "errors": 20,
-            "warnings": 80,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.contains("lint-rules show MD013")),
-            "should suggest lint-rules show for dominant rule (80%): {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_no_dominance_when_below_threshold() {
-        let c = ctx(HintSource::Lint);
-        // MD013 has 30 of 60 total (50% but only 30 absolute < 50 min).
-        let data = json!({
-            "files": [
-                {"file": "a.md", "rule_groups": [
-                    {"rule": "MD013", "count": 30, "shown": 3, "truncated": true,
-                     "severity": "warn", "autofixable": false,
-                     "violations": [{"line": 1, "column": 1, "message": "line too long"}]}
-                ]},
-                {"file": "b.md", "rule_groups": [
-                    {"rule": "HYALO001", "count": 30, "shown": 3, "truncated": true,
-                     "severity": "error", "autofixable": true,
-                     "violations": [{"line": 2, "column": 1, "message": "bare []"}]}
-                ]}
-            ],
-            "total": 60,
-            "rules_fired": 2,
-            "files_with_violations": 2,
-            "files_checked": 5,
-            "files_truncated": false,
-            "errors": 30,
-            "warnings": 30,
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("lint-rules show")),
-            "should not suggest lint-rules show when below dominance threshold: {hints:?}"
-        );
-    }
-
-    // --- UX-7: smart fix/dry-run hints ---
-
-    #[test]
-    fn lint_hints_not_fix_mode_suggests_preview_not_apply() {
-        // When not in fix mode, only preview should be suggested, not apply.
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files": [{"file": "test.md", "rule_groups": [
-                {"rule": "MD009", "count": 2, "shown": 2, "truncated": false,
-                 "severity": "warn", "autofixable": true,
-                 "violations": [{"line": 3, "column": 10, "message": "trailing spaces"}]}
-            ]}],
-            "total": 2,
-            "rules_fired": 1,
-            "files_with_violations": 1,
-            "files_checked": 1,
-            "files_truncated": false,
-            "errors": 0,
-            "warnings": 2,
-        });
-        let hints = generate_hints(&c, &data, None);
-        // Should have preview hint.
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("--fix --dry-run")),
-            "non-fix mode should suggest preview: {hints:?}"
-        );
-        // Should NOT suggest direct apply (user should preview first).
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.cmd.contains("lint --fix") && !h.cmd.contains("--dry-run")),
-            "non-fix mode should NOT suggest apply directly: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_hints_fix_mode_applied_no_fix_hints() {
-        // When fix was applied (not dry-run), no fix hints.
-        let mut c = ctx(HintSource::Lint);
-        c.lint_is_fix = true;
-        // dry_run defaults to false
-        let data = json!({
-            "files": [],
-            "total_fixed": 3,
-            "total_remaining": 0,
-            "total_conflicts": 0,
-            "rules_fired": 1,
-            "files_with_violations": 0,
-            "files_checked": 3,
-            "files_truncated": false,
-            "remaining_errors": 0,
-            "remaining_warnings": 0,
-            "dry_run": false,
-        });
-        let hints = generate_hints(&c, &data, None);
-        // Should NOT suggest any lint --fix hints since we already applied.
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("lint --fix")),
-            "after applying fixes, should not suggest fix again: {hints:?}"
-        );
-    }
-
-    // --- slow_query_hint ---
-
-    fn data_empty_array() -> serde_json::Value {
-        json!([])
-    }
-
-    /// Slow find with no index should emit the slow-query hint.
-    #[test]
-    fn slow_query_hint_fires_for_slow_find() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 1);
-        let h = slow_query_hint(&c);
-        assert!(h.is_some(), "expected slow-query hint");
-        let h = h.unwrap();
-        assert!(h.cmd == "hyalo create-index", "cmd: {}", h.cmd);
-        assert!(h.description.contains("ms"), "desc: {}", h.description);
-    }
-
-    /// Exactly at the threshold (not strictly greater) should NOT fire.
-    #[test]
-    fn slow_query_hint_does_not_fire_at_threshold() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS);
-        assert!(slow_query_hint(&c).is_none());
-    }
-
-    /// Fast query should not emit the hint.
-    #[test]
-    fn slow_query_hint_does_not_fire_when_fast() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = Some(50);
-        assert!(slow_query_hint(&c).is_none());
-    }
-
-    /// `--quiet` suppresses the slow-query hint even when slow.
-    #[test]
-    fn slow_query_hint_suppressed_by_quiet() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 100);
-        c.quiet = true;
-        assert!(slow_query_hint(&c).is_none());
-    }
-
-    /// Active index suppresses the slow-query hint even when slow.
-    #[test]
-    fn slow_query_hint_suppressed_when_has_index() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 100);
-        c.has_index = true;
-        assert!(slow_query_hint(&c).is_none());
-    }
-
-    /// Missing elapsed (`None`) means not yet measured — no hint.
-    #[test]
-    fn slow_query_hint_not_emitted_when_elapsed_none() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = None;
-        assert!(slow_query_hint(&c).is_none());
-    }
-
-    /// Ineligible source (e.g. Set) never emits slow-query hint.
-    #[test]
-    fn slow_query_hint_not_emitted_for_ineligible_source() {
-        let mut c = ctx(HintSource::Set);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 100);
-        assert!(slow_query_hint(&c).is_none());
-    }
-
-    /// All eligible sources should emit the hint when slow and no index.
-    #[test]
-    fn slow_query_hint_fires_for_all_eligible_sources() {
-        for source in [
-            HintSource::Find,
-            HintSource::Lint,
-            HintSource::Backlinks,
-            HintSource::PropertiesSummary,
-            HintSource::TagsSummary,
-            HintSource::Summary,
-            HintSource::Read,
-        ] {
-            let mut c = ctx(source);
-            c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 1);
-            assert!(
-                slow_query_hint(&c).is_some(),
-                "expected slow-query hint for source"
-            );
-        }
-    }
-
-    /// Slow-query hint appears in generate_hints output (via generate_hints_with_counters).
-    #[test]
-    fn slow_query_hint_surfaces_through_generate_hints() {
-        let mut c = ctx(HintSource::Find);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 1);
-        let hints = generate_hints(&c, &data_empty_array(), Some(0));
-        assert!(
-            hints.iter().any(|h| h.cmd == "hyalo create-index"),
-            "expected create-index hint: {hints:?}"
-        );
-    }
-
-    // --- large-vault summary hint ---
-
-    fn summary_data(files_total: u64) -> serde_json::Value {
-        json!({
-            "files": {"total": files_total, "by_directory": []},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [],
-            "tasks": {"total": 0, "done": 0},
-            "recent_files": []
-        })
-    }
-
-    /// Large vault (above threshold) with no index should emit the large-vault hint.
-    #[test]
-    fn large_vault_summary_hint_fires_when_over_threshold() {
-        let c = ctx(HintSource::Summary);
-        let data = summary_data(LARGE_VAULT_FILE_COUNT + 1);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd == "hyalo create-index" && h.description.contains("files")),
-            "expected large-vault hint: {hints:?}"
-        );
-    }
-
-    /// Exactly at the threshold (not strictly greater) should NOT fire.
-    #[test]
-    fn large_vault_summary_hint_does_not_fire_at_threshold() {
-        let c = ctx(HintSource::Summary);
-        let data = summary_data(LARGE_VAULT_FILE_COUNT);
-        let hints = generate_hints(&c, &data, None);
-        // The hint should not appear.
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.cmd == "hyalo create-index" && h.description.contains("files")),
-            "unexpected large-vault hint at threshold: {hints:?}"
-        );
-    }
-
-    /// Small vault should not emit the large-vault hint.
-    #[test]
-    fn large_vault_summary_hint_not_fired_for_small_vault() {
-        let c = ctx(HintSource::Summary);
-        let data = summary_data(10);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.cmd == "hyalo create-index" && h.description.contains("files")),
-            "unexpected large-vault hint for small vault: {hints:?}"
-        );
-    }
-
-    /// Active index suppresses the large-vault hint.
-    #[test]
-    fn large_vault_summary_hint_suppressed_when_has_index() {
-        let mut c = ctx(HintSource::Summary);
-        c.has_index = true;
-        let data = summary_data(LARGE_VAULT_FILE_COUNT + 100);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.cmd == "hyalo create-index" && h.description.contains("files")),
-            "large-vault hint should be suppressed with active index: {hints:?}"
-        );
-    }
-
-    /// `--quiet` suppresses the large-vault hint (parity with slow-query hint).
-    #[test]
-    fn large_vault_summary_hint_suppressed_by_quiet() {
-        let mut c = ctx(HintSource::Summary);
-        c.quiet = true;
-        let data = summary_data(LARGE_VAULT_FILE_COUNT + 100);
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.cmd == "hyalo create-index" && h.description.contains("files")),
-            "large-vault hint should be suppressed by --quiet: {hints:?}"
-        );
-    }
-
-    /// When both slow-query and large-vault conditions fire, only one
-    /// `create-index` hint should appear in the envelope (dedupe by `cmd`).
-    #[test]
-    fn create_index_hint_deduped_when_both_conditions_fire() {
-        let mut c = ctx(HintSource::Summary);
-        c.elapsed_ms = Some(SLOW_QUERY_THRESHOLD_MS + 1);
-        let data = summary_data(LARGE_VAULT_FILE_COUNT + 100);
-        let hints = generate_hints(&c, &data, None);
-        let n = hints
-            .iter()
-            .filter(|h| h.cmd == "hyalo create-index")
-            .count();
-        assert_eq!(
-            n, 1,
-            "expected exactly one create-index hint, got {n}: {hints:?}"
-        );
-    }
-
-    /// NEW-1 regression: create-index hint must appear even when orphans +
-    /// broken-links + links-fix hints would otherwise consume all MAX_HINTS
-    /// slots. On real large vaults (MDN: 4245 orphans, 49933 broken links)
-    /// the health hints used to crowd out the index hint entirely.
-    #[test]
-    fn create_index_hint_visible_even_when_health_hints_fill_cap() {
-        let c = ctx(HintSource::Summary);
-        // A large vault with both orphans and broken links to generate many hints.
-        let data = json!({
-            "files": {"total": LARGE_VAULT_FILE_COUNT + 14000, "by_directory": []},
-            "orphans": 4245u64,
-            "dead_ends": 100u64,
-            "links": {"total": 100_000u64, "broken": 49_933u64},
-            "properties": [],
-            "tags": {"tags": [], "total": 0},
-            "status": [],
-            "tasks": {"total": 0, "done": 0},
-            "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        // Total hints must not exceed MAX_HINTS.
-        assert!(
-            hints.len() <= MAX_HINTS,
-            "hints exceeded MAX_HINTS={MAX_HINTS}: {hints:?}"
-        );
-        // The create-index hint must be present despite the health-hint pressure.
-        assert!(
-            hints.iter().any(|h| h.cmd == "hyalo create-index"),
-            "create-index hint missing; hints: {hints:?}"
-        );
-        // create-index should be the first hint (highest priority).
-        assert_eq!(
-            hints[0].cmd, "hyalo create-index",
-            "create-index should be first hint; got: {hints:?}"
-        );
-    }
-
-    // --- okf hints (iter-177) ---
-
-    #[test]
-    fn okf_index_dry_run_drift_suggests_apply_then_validate() {
-        let c = ctx(HintSource::OkfIndex);
-        let data = json!({ "apply": false, "changed": 2 });
-        let hints = generate_hints(&c, &data, None);
-        assert_eq!(hints[0].cmd, "hyalo okf index --apply");
-        assert_eq!(hints[1].cmd, "hyalo lint --profile okf");
-    }
-
-    #[test]
-    fn okf_index_clean_dry_run_only_validate() {
-        let c = ctx(HintSource::OkfIndex);
-        let data = json!({ "apply": false, "changed": 0 });
-        let hints = generate_hints(&c, &data, None);
-        assert_eq!(hints.len(), 1);
-        assert_eq!(hints[0].cmd, "hyalo lint --profile okf");
-    }
-
-    #[test]
-    fn okf_index_apply_omits_apply_hint() {
-        let c = ctx(HintSource::OkfIndex);
-        // Even with changes, an --apply run should not re-suggest applying.
-        let data = json!({ "apply": true, "changed": 3 });
-        let hints = generate_hints(&c, &data, None);
-        assert_eq!(hints.len(), 1);
-        assert_eq!(hints[0].cmd, "hyalo lint --profile okf");
-    }
-
-    #[test]
-    fn okf_validate_hint_drops_redundant_profile_flag_when_active() {
-        let mut c = ctx(HintSource::OkfLog);
-        c.okf_profile_active = true;
-        let hints = generate_hints(&c, &json!({}), None);
-        assert_eq!(hints[0].cmd, "hyalo lint");
-    }
-
-    // --- iter-180: hint trust ---
-
-    #[test]
-    fn slow_query_create_index_hint_carries_dir() {
-        // BUG-7: the slow-query create-index hint dropped an explicit --dir.
-        let mut c = ctx_with_dir(HintSource::Find, "/my/vault");
-        c.elapsed_ms = Some(1062);
-        let items = vec![make_find_item("a.md", None, &[])];
-        let hints = generate_hints(&c, &json!(items), None);
-        let ci = hints
-            .iter()
-            .find(|h| h.cmd.starts_with("hyalo create-index"))
-            .expect("slow-query create-index hint expected");
-        assert_eq!(ci.cmd, "hyalo create-index --dir /my/vault");
-        // BUG-7: no dangling colon on the description.
-        assert!(
-            !ci.description.ends_with(':'),
-            "description should not end with a colon: {:?}",
-            ci.description
-        );
-    }
-
-    #[test]
-    fn summary_large_vault_create_index_hint_carries_dir() {
-        let c = ctx_with_dir(HintSource::Summary, "/vault");
-        let data = json!({
-            "files": {"total": 501, "by_directory": []},
-            "properties": [], "tags": {"tags": [], "total": 0},
-            "status": [], "tasks": {"total": 0, "done": 0}, "recent_files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        let ci = hints
-            .iter()
-            .find(|h| h.cmd.starts_with("hyalo create-index"))
-            .expect("large-vault create-index hint expected");
-        assert_eq!(ci.cmd, "hyalo create-index --dir /vault");
-        assert!(!ci.description.ends_with(':'));
-    }
-
-    #[test]
-    fn find_orphan_show_all_preserves_orphan_filter() {
-        // BUG-8: "Show all N results" on a --orphan query must keep --orphan.
-        let mut c = ctx(HintSource::Find);
-        c.orphan_filter = true;
-        let items: Vec<serde_json::Value> = (0..50)
-            .map(|i| make_find_item(&format!("{i}.md"), None, &[]))
-            .collect();
-        // total (79) > shown (50) → truncated, so the show-all hint fires.
-        let hints = generate_hints(&c, &json!(items), Some(79));
-        let show_all = hints
-            .iter()
-            .find(|h| h.description.starts_with("Show all"))
-            .expect("show-all hint expected");
-        assert!(
-            show_all.cmd.contains("--orphan"),
-            "show-all hint must preserve --orphan: {}",
-            show_all.cmd
-        );
-        assert!(show_all.cmd.contains("--limit 0"));
-    }
-
-    #[test]
-    fn find_orphan_narrow_by_tag_composes_and_drops_stale_count() {
-        // BUG-8: the narrow-by-tag hint on a --orphan query must keep --orphan
-        // and, when the result set was truncated, drop the (misleading) count.
-        let mut c = ctx(HintSource::Find);
-        c.orphan_filter = true;
-        let items: Vec<serde_json::Value> = (0..50)
-            .map(|i| make_find_item(&format!("{i}.md"), None, &["iteration"]))
-            .collect();
-        let hints = generate_hints(&c, &json!(items), Some(79));
-        let narrow = hints
-            .iter()
-            .find(|h| h.description.starts_with("Narrow by tag"))
-            .expect("narrow-by-tag hint expected");
-        assert!(
-            narrow.cmd.contains("--orphan") && narrow.cmd.contains("--tag iteration"),
-            "narrow hint must compose --orphan with the new tag: {}",
-            narrow.cmd
-        );
-        // Truncated set → no parenthetical count in the description.
-        assert!(
-            !narrow.description.contains('('),
-            "count must be dropped when truncated: {:?}",
-            narrow.description
-        );
-    }
-
-    #[test]
-    fn find_narrow_by_tag_keeps_count_when_not_truncated() {
-        let c = ctx(HintSource::Find);
-        let items: Vec<serde_json::Value> = (0..6)
-            .map(|i| make_find_item(&format!("{i}.md"), None, &["iteration"]))
-            .collect();
-        // total == shown → not truncated → count is accurate and shown.
-        let hints = generate_hints(&c, &json!(items), Some(6));
-        let narrow = hints
-            .iter()
-            .find(|h| h.description.starts_with("Narrow by tag"))
-            .expect("narrow-by-tag hint expected");
-        assert!(
-            narrow.description.contains("(6 files)"),
-            "count should be shown when not truncated: {:?}",
-            narrow.description
-        );
-    }
-
-    /// iter-210 / UX-5: the snapshot path is the longest token a hint carries
-    /// and it is repeated on every derived query. Inside the working directory
-    /// it renders relative — shorter, and still runnable verbatim.
-    #[test]
-    fn index_path_renders_relative_to_the_working_directory() {
-        let cwd = std::env::current_dir().expect("cwd");
-        let inside = cwd.join("sub").join(".hyalo-index");
-        assert_eq!(shorten_index_path_for_hint(&inside), "sub/.hyalo-index");
-    }
-
-    /// A path outside the working directory has no shorter runnable spelling,
-    /// so it stays absolute rather than turning into a `../..` chain.
-    #[test]
-    fn index_path_outside_the_working_directory_stays_absolute() {
-        let outside = std::path::Path::new(if cfg!(windows) {
-            r"C:\definitely\elsewhere\.hyalo-index"
-        } else {
-            "/definitely/elsewhere/.hyalo-index"
-        });
-        assert_eq!(
-            shorten_index_path_for_hint(outside),
-            outside.display().to_string()
-        );
-    }
-
-    #[test]
-    fn find_index_file_preserved_in_derived_hints() {
-        // BUG-7 audit: --index-file was a dropped flag.
-        let mut c = ctx(HintSource::Find);
-        c.find_index = FindIndexHint::File("sub/.hyalo-index".to_owned());
-        let items: Vec<serde_json::Value> = (0..50)
-            .map(|i| make_find_item(&format!("{i}.md"), None, &[]))
-            .collect();
-        let hints = generate_hints(&c, &json!(items), Some(200));
-        let show_all = hints
-            .iter()
-            .find(|h| h.description.starts_with("Show all"))
-            .expect("show-all hint expected");
-        assert!(
-            show_all.cmd.contains("--index-file sub/.hyalo-index"),
-            "derived hint must preserve --index-file: {}",
-            show_all.cmd
-        );
-    }
-
-    #[test]
-    fn summary_schema_hint_relabeled_and_targets_schema_rule() {
-        // BUG-9: relabel "Lint" → "Schema" and target `lint --rule SCHEMA`.
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 10, "by_directory": []},
-            "properties": [], "tags": {"tags": [], "total": 0},
-            "status": [], "tasks": {"total": 0, "done": 0}, "recent_files": [],
-            "schema": {"errors": 5, "warnings": 12, "files_with_violations": 8}
-        });
-        let hints = generate_hints(&c, &data, None);
-        let schema_hint = hints
-            .iter()
-            .find(|h| h.description.starts_with("Schema:"))
-            .expect("schema hint expected");
-        assert_eq!(schema_hint.description, "Schema: 5 errors, 12 warnings");
-        assert_eq!(schema_hint.cmd, "hyalo lint --rule SCHEMA");
-        assert!(
-            !hints.iter().any(|h| h.description.starts_with("Lint:")),
-            "should not use the old \"Lint:\" label: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_fix_apply_suppresses_stale_show_all_hint() {
-        let mut c = ctx(HintSource::Lint);
-        c.lint_is_fix = true; // apply (not dry-run)
-        let data = json!({
-            "files_truncated": true,
-            "files_with_issues": 42,
-            "files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            !hints.iter().any(|h| h.description.contains("Show all")),
-            "post-apply output must not carry the stale show-all hint: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn lint_readonly_keeps_show_all_hint() {
-        let c = ctx(HintSource::Lint);
-        let data = json!({
-            "files_truncated": true,
-            "files_with_issues": 42,
-            "files": []
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.description.contains("Show all")),
-            "read-only lint should keep the show-all hint: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn summary_site_url_heuristic_replaces_links_fix() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 14000, "by_directory": []},
-            "properties": [], "tags": {"tags": [], "total": 0},
-            "status": [], "tasks": {"total": 0, "done": 0}, "recent_files": [],
-            "orphans": 0, "dead_ends": 0,
-            "links": {"total": 49935, "broken": 49933}
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints
-                .iter()
-                .any(|h| h.cmd.is_empty() && h.description.contains("--site-prefix")),
-            "site-URL diagnostic expected: {hints:?}"
-        );
-        assert!(
-            !hints.iter().any(|h| h.cmd.contains("links fix")),
-            "links fix should be suppressed when links look like site URLs: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn summary_normal_broken_links_still_offers_links_fix() {
-        let c = ctx(HintSource::Summary);
-        let data = json!({
-            "files": {"total": 100, "by_directory": []},
-            "properties": [], "tags": {"tags": [], "total": 0},
-            "status": [], "tasks": {"total": 0, "done": 0}, "recent_files": [],
-            "orphans": 0, "dead_ends": 0,
-            "links": {"total": 400, "broken": 12}
-        });
-        let hints = generate_hints(&c, &data, None);
-        assert!(
-            hints.iter().any(|h| h.cmd.contains("links fix")),
-            "a handful of broken links should still offer links fix: {hints:?}"
-        );
-    }
-
-    #[test]
-    fn find_show_all_and_narrow_hints_preserve_active_sort() {
-        // Milder variant of BUG-8: a truncated, explicitly sorted query's
-        // derived hints (show-all, narrow-by-tag) must keep --sort/--reverse,
-        // else "show all" / "narrow by tag" silently reverts to default
-        // ordering instead of reproducing the query that produced them.
-        let mut c = ctx(HintSource::Find);
-        c.sort = Some("modified".to_owned());
-        c.reverse = true;
-        let items: Vec<serde_json::Value> = (0..50)
-            .map(|i| make_find_item(&format!("{i}.md"), None, &["iteration"]))
-            .collect();
-        let hints = generate_hints(&c, &json!(items), Some(79));
-
-        let show_all = hints
-            .iter()
-            .find(|h| h.description.starts_with("Show all"))
-            .expect("show-all hint expected");
-        assert!(
-            show_all.cmd.contains("--sort modified") && show_all.cmd.contains("--reverse"),
-            "show-all hint must preserve --sort/--reverse: {}",
-            show_all.cmd
-        );
-
-        let narrow = hints
-            .iter()
-            .find(|h| h.description.starts_with("Narrow by tag"))
-            .expect("narrow-by-tag hint expected");
-        assert!(
-            narrow.cmd.contains("--sort modified") && narrow.cmd.contains("--reverse"),
-            "narrow-by-tag hint must preserve --sort/--reverse: {}",
-            narrow.cmd
-        );
-
-        // Because ctx.sort is Some, the literal "Sort by most recently
-        // modified" suggestion must not also fire (it only applies when no
-        // sort is active).
-        assert!(
-            !hints
-                .iter()
-                .any(|h| h.description == "Sort by most recently modified"),
-            "should not suggest sorting when a sort is already active: {hints:?}"
-        );
-    }
-}
+mod tests;
